% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AminoAcidSequenceSearch.R
\name{AAalign}
\alias{AAalign}
\title{Generate an amino acid alignment.}
\usage{
AAalign(alleles, positions)
}
\arguments{
\item{alleles}{A vector of un-prefixed HLA locus names.}

\item{positions}{either a vector of amino acid positions, against which all loci will be aligned, or a list of vectors of AA positions, exactly one vector for each allele, against which each corresponding allele will be aligned.}
}
\value{
A data frame of allele names and the corresponding peptide sequence for each amino acid position. an error message is returned if input loci is not available in the ANHIG/IMGTHLA Github Repository.
}
\description{
Generates an amino acid alignment table for user-specified HLA alleles at user-specified amino acid positions.
}
\examples{
\dontrun{
AAalign(c("DRB1*01:01","DQB1*02:01","DPB1*01:01"),c(1,2,3,7,8,9,13,14,15))
AAalign(c("DQA1*01:01","DQB1*05:01","DPB1*01:01","DPA1*01:03"),list(32:58,33:59,31:57,29:55))
}
}
