% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{allele_id}
\alias{allele_id}
\title{Access allele IDs}
\usage{
allele_id(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of allele IDs.
}
\description{
Access allele IDs
}
\examples{
\dontrun{
dpb1 <- parse_hla_alleles(read_hla_xml(), "DPB1")
allele_id(dpb1[is_complete(dpb1), ])
}
}

