% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{allele_name}
\alias{allele_name}
\title{Access allele names}
\usage{
allele_name(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of allele names.
}
\description{
Access allele names
}
\examples{
\dontrun{
dpb1 <- parse_hla_alleles(read_hla_xml(), "DPB1")
allele_name(dpb1[is_complete(dpb1), ])
}
}

