% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_tables.R
\name{allele_table}
\alias{allele_table}
\title{Class: allele_tbl}
\source{
\href{https://www.hla.alleles.org/wmda}{hla.alleles.org}.
}
\usage{
allele_table(db_path = getOption("hlatools.local_repos"), remote = FALSE)
}
\arguments{
\item{db_path}{<\link{character}>; location of local IPD-IMGT/HLA repository.}

\item{remote}{<\link{logical}>; if \code{TRUE} pull data from \href{https://www.hla.alleles.org/wmda}{hla.alleles.org},
if \code{FALSE} retrieve data from \code{db_path}.}
}
\value{
A \link{tibble} with the fields:
\itemize{
\item "gene": The HLA gene.
\item "allele_name": HLA Allele name.
\item "date_assigned": Date assigned.
\item "date_deleted": Date deleted, if the name has now been abandoned or \code{NA}.
\item "identical_to":  Allele that the deleted allele was shown to be identical to.
\item "reason_for_deletion": Reason for the Allele to be deleted.
}
}
\description{
Constructor for a <\code{allele_tbl}> object.
}
\examples{
\dontrun{
a_tbl <- allele_table()
}
}
\seealso{
\link{nmdp_table}, \link{g_table}
}
