% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alleleListHistory}
\alias{alleleListHistory}
\title{Identifies HLA loci that have amino acid, cDNA, or gDNA alignments.}
\format{
A large list of 2 elements that identify the table of allele names for each IPD-IMGT/HLA reference Database release and the source release
\itemize{
\item(Version: the Date and IPD-IMGT HLA Release version under which the alleleListHistory object was made)
\item(AlleleListHistory: a dataframe version of the alleleListHistory.txt file)
}
}
\source{
https://raw.githubusercontent.com/ANHIG/IMGTHLA/Latest/Allelelist_history.txt
}
\usage{
data(alleleListHistory)
}
\description{
A large list object of two elements. The first element contains Allele List version information. The second element contains all releases of allele lists on the IMGT/HLA database.
This object is built by the UpdateAlleleListHistory() function in the HLAtools package.
}
\keyword{datasets}
