% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{ancestry}
\alias{ancestry}
\title{Access ancestry status}
\usage{
ancestry(x, ...)
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} or \linkS4class{HLAAllele} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of colon-separated sample ancestries.
}
\description{
Access ancestry status
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
ancestry(x[is_complete(x)])
}
}
