% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDstrat.R
\name{BDstrat}
\alias{BDstrat}
\title{Allele stratification for BIGDAWG datasets}
\usage{
BDstrat(dataset, allele, warnBelow = 21)
}
\arguments{
\item{dataset}{A BIGDAWG-formatted data.frame or a path to a BIGDAWG-formatted tab-delimited text file.}

\item{allele}{A single allele in the \if{html}{\out{<locus>}}*\if{html}{\out{<allele name>}} format (e.g., "A\*01:01:01:01" for HLA alleles or "KIR2DL1\*001" for KIR alleles).}

\item{warnBelow}{An integer that defines a low number of subjects in a stratum, generating a warning message. The default value is 21.}
}
\value{
A list-object of two BIGDAWG-formatted data frames titled dataset$\code{allele-positive} and dataset$\code{allele-negative}. The positive list element includes all subjects with a specified allele, and the negative list element includes all subjects without that specified allele.
}
\description{
Divides a BIGDAWG-formatted dataset into two subsets (strata) that do and do not include a specific allele.
}
\examples{
\dontrun{
HLA_data.strat <- BDstrat(BIGDAWG::HLA_data,"DRB1*15:01:01:01",Run.Tests = "L",HLA=TRUE)
for(i in 1:2) {BIGDAWG(HLA_data.strat[[i]],HLA = TRUE,Run.Tests = "L")}
}

}
