% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGazeteer.R
\name{buildGazeteer}
\alias{buildGazeteer}
\title{Build lists of genes}
\usage{
buildGazeteer(version = getLatestVersion())
}
\arguments{
\item{version}{A string identifying of the desired IPD-IMGT/HLA Database release version to which the alleles should be updated. The default value is most recent IPD-IMGT/HLA Database release version.}
}
\value{
A list object of vectors identifying those loci in the IPD-IMGT/HLA Database for which amino acid (prot), cDNA (nuc), and gDNA (gen) alignments are available. In addition, subsets of these loci identified as pseudogenes (pseudo) and gene fragments (frag) for which nucleotide and genomic alignments are and are not available (the latter denoted as 'noNuc' or 'noGen') are also identified.
}
\description{
Queries the ANHIG/IMGTHLA GitHub Repository to identify HLA genes that do and do not have amino acid, cDNA, and genomic (gDNA) alignments.
}
\note{
Pseudogenes and gene fragments are defined by the IPD-IMGT/HLA Database at https://hla.alleles.org/genes/index.html, and are described in the HLAtools::IMGTHLAGeneTypes data object.

This function requires internet access to function.
}
\examples{
\dontrun{
HLAgazeteer <- buildGazeteer("3.34.0")
}

}
