% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenomicReferenceGenes.R
\name{buildGenomicReferences}
\alias{buildGenomicReferences}
\title{Generate a Table of Genomic Reference Alleles and Characteristics of HLA Genes Defined by the IPD-IMGT/HLA Database}
\usage{
buildGenomicReferences()
}
\value{
A list object of two elements -- GenomicReferences and Version.
}
\description{
buildGenomicReferences() extracts the table of genomic HLA reference alleles at 'www.ebi.ac.uk/ipd/imgt/hla/about/help/genomics/' and generates the IMGTGenomicReferences data object.
}
\note{
This function requires internet access to function.

For internal HLAtools use.
}
\references{
Barker et al. NAR (2025) 54:D1152–D1158 https://doi.org/10.1093/nar/gkaf1218
}
