% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryFunctions.R
\name{buildHTexceptions}
\alias{buildHTexceptions}
\title{Builds 'HTexceptions', a list object of sub-lists of vectors that define exceptions for specific use cases, saves it as an Rdata file in the 'HLAtools/data' folder, and loads the HTexceptions object into the R environment.}
\usage{
buildHTexceptions(HTpath = paste(path.package("HLAtools"), "/data", sep = ""))
}
\arguments{
\item{HTpath}{The path to the location where the 'HTexceptions.rda' object should be stored. The default value is the path to the /data folder of the installed HLAtools package.}
}
\value{
The date stamp of the newly created HTexceptions.rda file.
}
\description{
Builds 'HTexceptions', a list object of sub-lists of vectors that define exceptions for specific use cases, saves it as an Rdata file in the 'HLAtools/data' folder, and loads the HTexceptions object into the R environment.
}
\note{
For internal HLAtools use.
}
\examples{
\dontrun{
buildHTexceptions()
}

}
