% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDNASequenceSearch.R
\name{cDNAalign}
\alias{cDNAalign}
\title{Generate a codon DNA sequence alignment.}
\usage{
cDNAalign(alleles, positions)
}
\arguments{
\item{alleles}{A vector of un-prefixed HLA locus names.}

\item{positions}{Either a vector of nucleotide positions, against which all loci will be aligned, or a list of vectors of nucleotide positions, exactly one vector for each allele, against which each corresponding allele will be aligned.}
}
\value{
A data frame of allele names and the corresponding nucleotide sequences for each desired nucleotide position. an error message is returned if input loci is not available in the ANHIG/IMGTHLA Github Repository.
}
\description{
Generates a cDNA alignment table for user-specified HLA alleles at user-specified codon positions.
}
\examples{
\dontrun{
cDNAalign(c("DRB1*01:01","DQB1*02:01","DPB1*01:01"),c(1,2,3,7,8,9,13,14,15))
cDNAalign(c("DQA1*01:01:01:01","DQB1*05:01:01:01","DPB1*01:01:01:01"),list(32:58,33:59,31:57))
}
}
