% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_remotes}
\alias{check_remotes}
\title{Set remote data checking on of off}
\usage{
check_remotes(what)
}
\arguments{
\item{what}{\code{TRUE} or \code{FALSE}}
}
\description{
Set remote data checking on of off
}
\keyword{internal}
