% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryFunctions.R
\name{checkAlignType}
\alias{checkAlignType}
\title{Evaluates 'alignType' values to ensure that only "prot", "nuc", "codon" and "gen" are passed to downstream functions. If any other values are entered, a message describing excluded values is generated.}
\usage{
checkAlignType(alignType)
}
\arguments{
\item{alignType}{A vector of character values specifying sequence alignment types to be used for a function.}
}
\value{
A vector that includes only allowed 'alignType' values.
}
\description{
Evaluates 'alignType' values to ensure that only "prot", "nuc", "codon" and "gen" are passed to downstream functions. If any other values are entered, a message describing excluded values is generated.
}
\examples{
\dontrun{
checkAlignType(c("nuc","prot","gDNA")) 
}

}
