% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryFunctions.R
\name{checkSource}
\alias{checkSource}
\title{Ensure that Source Values are Valid}
\usage{
checkSource(source)
}
\arguments{
\item{source}{A vector of character values specifying sequence alignment file sources to be used for a function.}
}
\value{
A character vector that includes only allowed 'source' values.
}
\description{
Evaluates 'source' values to ensure that only "AA", "cDNA", and "gDNA" are passed to downstream functions. If any other values are entered, a message describing excluded values is generated.
}
\examples{
checkSource(c("AA","cDNA","codon")) 

}
