% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{clone_IMGTHLA}
\alias{clone_IMGTHLA}
\alias{pull_IMGTHLA}
\alias{check_IMGTHLA}
\title{Clone or pull the \href{https://github.com/ANHIG/IMGTHLA}{ANHIG/IMGTHLA} repository}
\usage{
clone_IMGTHLA(db_path = getOption("hlatools.local_repos"), branch = NULL,
  progress = TRUE)

pull_IMGTHLA(db_path = getOption("hlatools.local_repos"))

check_IMGTHLA(db_path = getOption("hlatools.local_repos"))
}
\arguments{
\item{db_path}{Local directory to clone into. Defaults to \file{~/local/db/IMGTHLA}.}

\item{branch}{The name of the branch to checkout. Defaults to the remote's default branch.}

\item{progress}{Show progress. Default is \code{TRUE}.}
}
\value{
A \linkS4class{git_repository} object.

\link{character}; a diagnostic message.
}
\description{
Clone or pull the \href{https://github.com/ANHIG/IMGTHLA}{ANHIG/IMGTHLA} repository
}
\examples{
\dontrun{
check_IMGTHLA()
repo <- clone_IMGTHLA()
repo <- pull_IMGTHLA()
}
}
