% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_composite.R
\name{composite_pipeline}
\alias{composite_pipeline}
\title{Full pipeline to predict best HLA allele from multiple genotypers
using pre-trained decision tree}
\usage{
composite_pipeline(df, digits = 4)
}
\arguments{
\item{df}{Data frame of HLA alleles genotypers as read by the read_* set of functions}

\item{digits}{Which allele resolution to compile}
}
\value{
data frame
}
\description{
Full pipeline to predict best HLA allele from multiple genotypers
using pre-trained decision tree
}
\examples{
# No examples
}
