% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDtoPyPop.R
\name{convertAny}
\alias{convertAny}
\title{Convert Values Across an Entire Data Frame or Vector

Converts all instances of a value in a data frame or vector to a specified value.}
\usage{
convertAny(dataset, change.from = NA, change.to = "****")
}
\arguments{
\item{dataset}{A data frame or vector of values.}

\item{change.from}{The value present in the data frame or vector that should be changed. The default value is NA.}

\item{change.to}{The value to which 'change.from' should be changed to. The default value is "****".}
}
\value{
A data frame or vector in which 'change.from' has been converted to 'change.to'.
}
\description{
Convert Values Across an Entire Data Frame or Vector

Converts all instances of a value in a data frame or vector to a specified value.
}
\examples{
dataset <- convertAny(sHLAdata)

}
