% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThirdPartyFunctions.R
\name{countSpaces}
\alias{countSpaces}
\title{Count the Spaces in a Character String}
\source{
https://www.reddit.com/r/rstats/comments/2th8ic/function_to_count_the_number_of_white_spaces/
}
\usage{
countSpaces(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A vector of numeric values identifying the number of instances of spaces (' ') in x, and the number of consecutive spaces in an instance.
}
\description{
Counts the number of spaces in a character string.
}
\examples{
countSpaces("abc def")
countSpaces("abc def  ghi   jkl mno")

}
\author{
Josh Bredeweg, Reddit user jbraids1421
}
