% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{cwd_status}
\alias{cwd_status}
\title{Access CWD status}
\usage{
cwd_status(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of CWD status codes.
}
\description{
Access CWD status
}
\examples{
##
}

