% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{db_version}
\alias{db_version}
\title{Get IMGT/HLA database version}
\usage{
db_version(x, ...)
}
\arguments{
\item{x}{A \code{\link{HLAGene}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector.
}
\description{
Get IMGT/HLA database version
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
db_version(x)
}
}

