% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drb345_data}
\alias{drb345_data}
\title{Counts of reads aligned to the human DRB loci for 346 subjects}
\format{
A data frame with 346 rows and 5 columns:
\describe{
  \item{sample}{A unique sample name}
  \item{DRB1, DRB3, DRB4, DRB5}{Raw counts of reads aligned to each locus}
}
}
\source{
https://github.com/BenSolomon/hla_benchmark
}
\usage{
drb345_data
}
\description{
Counts of reads aligned to the human DRB loci for 346 subjects
}
\keyword{datasets}
