% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_drb345.R
\name{drb345_pipeline}
\alias{drb345_pipeline}
\title{Full pipeline to filter HLA genotypes based on DRB345 kNN}
\usage{
drb345_pipeline(arcasHLA_alleles, arcasHLA_reads)
}
\arguments{
\item{arcasHLA_alleles}{Data frame of arcasHLA genotypes as read by read_arcasHLA}

\item{arcasHLA_reads}{Data frame of arcasHLA read mapping statistics as read by read_arcasHLA_reads}
}
\value{
A dataframe identical to that provided to arcasHLA_alleles with exception 
of DRB345 loci, which are filtered based on copy numbers predicted by predict_drb345
}
\description{
Full pipeline to filter HLA genotypes based on DRB345 kNN
}
\examples{
# No example
}
