% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{ethnicity}
\alias{ethnicity}
\title{Access ethnicity status}
\usage{
ethnicity(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of colon-separated sample ethnicities.
}
\description{
Access ethnicity status
}
\examples{
\dontrun{
dpb1 <- parse_hla_alleles(read_hla_xml(), "DPB1")
etnicity(dpb1[is_complete(dpb1), ])
}
}

