% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionValidation.R
\name{expandVersion}
\alias{expandVersion}
\title{Add 'Dot' Delimiters to a Numeric Release Version}
\usage{
expandVersion(ver)
}
\arguments{
\item{ver}{A four character/numeral IPD-IMGT/HLA Database Release Version name (e.g., '3540' or 3450).}
}
\value{
A character vector of the expanded IPD-IMGT/HLA Database Release Version name.
}
\description{
Adds the '.' delimiters to a 'squashed' IPD-IMGT/HLA Database Release Version name
}
\note{
For internal HLAtools use.

This function assumes single-character-length first and third fields of IPD-IMGT/HLA Database Release Version names.
}
\examples{
expandVersion(3450)

}
