% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{features}
\alias{features}
\alias{features<-}
\title{Get or set feature data}
\usage{
features(x, ...)

features(x, ...) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}

\item{value}{A \code{HLAFeatureList} object.}
}
\value{
A HLAFeatureList object.
}
\description{
Get or set feature data
}
\examples{
##
}

