% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fragmentFeatureNames}
\alias{fragmentFeatureNames}
\title{Gene Features of HLA Pseduogenes and Gene Fragments}
\format{
A list of 15 elements that identify the gene features for each locus wih a genomic alignment and annotate non-standard features
}
\source{
https://hla.alleles.org/genes/index.html

https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments
}
\usage{
data(fragmentFeatureNames)
}
\description{
A list object of 15 elements. Each of the 15 elements corresponds to one of the 15 HLA pseudogenes and gene fragments, and contains two items. The first item identifies the gene features for that locus. The second item contains an annotation detailing information regarding non-standard gene structures.
This object is built by the ffN() function.
}
\keyword{datasets}
