% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_tables.R
\name{g_table}
\alias{g_table}
\title{Class: g_tbl}
\source{
\href{https://www.hla.alleles.org/nomenclature/g_groups.html}{hla.alleles.org}.
}
\usage{
g_table(db_path = getOption("hlatools.local_repos"), remote = FALSE)
}
\arguments{
\item{db_path}{<\link{character}>; location of local IPD-IMGT/HLA repository.}

\item{remote}{<\link{logical}>; if \code{TRUE} pull data from \href{https://www.hla.alleles.org/wmda}{hla.alleles.org},
if \code{FALSE} retrieve data from \code{db_path}.}
}
\value{
A \link{tibble} with the fields:
\itemize{
\item "gene": The HLA gene.
\item "code": The G group name.
\item "subtype": The '/'-separated subtypes into which a G Code expands.
}
}
\description{
Constructor for a <\code{g_tbl}> object.
}
\examples{
\dontrun{
g_codes <- g_table()
}
}
\seealso{
\link{nmdp_table}, \link{allele_table}
}
