% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDNASequenceSearch.R
\name{gDNAsearch}
\alias{gDNAsearch}
\title{Search genomic HLA alignments for specific alleles and nucleotide positions.}
\usage{
gDNAsearch(allelename, positions, prefix = TRUE, sep = "~")
}
\arguments{
\item{allelename}{A full-length or 2-field HLA allele name, excluding the "HLA-" prefix, of an HLA allele with a genomic alignment.}

\item{positions}{A vector of nucleotide positions (e.g., 'c(-17,1,22,130)' or '25:35') in the pertinent alignment; indel positions are identified using decimals. E.g., the first indel between positions 26 and 27 is named 26.1, and the second indel between 26 and 27 is 26.2, etc.}

\item{prefix}{A logical that indicates if the position number(s) should be included in the returned string.}

\item{sep}{The value that will be used to separate the sequences for each position. No value can be specified (sep=""). If prefix=FALSE and sep="", a sequence block is returned. The default value of 'sep' is "~". If a trimmed (two-field version of a 3- or 4- field name) is provided, the sequence of the lowest-numbered full-length name is returned.}
}
\value{
A character string containing the corresponding nucleotide sequence for each position in 'positions' for 'allelename'. A nucleotide sequence is not returned if 'allelename' is not not found in the pertinent alignment. A position will be empty if 'allelename' does not have a nucleotide at the specified position.
}
\description{
Generates a character string of nucleotide variants at specified positions in the genomic (gDNA) sequence of an HLA allele, using IPD-IMGT/HLA genomic alignments in the HLAtools.data package.
}
\examples{
\dontrun{
gDNAsearch("DRB1*15:07:01",11:22)
gDNAsearch("DRB1*15:07:01",11:22,prefix = FALSE,sep="")
gDNAsearch("DRB1*15:07",11:22)
gDNAsearch("DRB1*15:07:01",c(321,321.1,322),prefix = FALSE,sep="")
}
}
