% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGazeteer.R
\name{getAlignmentNames}
\alias{getAlignmentNames}
\title{Retrieve alignment filenames for HLA genes}
\usage{
getAlignmentNames(URL)
}
\arguments{
\item{URL}{A URL identifying of the desired IPD-IMGT/HLA Database release version from which the alignment filenames should be retrieved.}
}
\value{
A character vector of all of the filenames.
}
\description{
Retrieves the filenames of the protein, nucleotide and genomic alignments available a specific branch of the IMGTHLA GitHub Repository
}
\examples{
\dontrun{
getAlignmentNames("https://github.com/ANHIG/IMGTHLA/tree/3240/alignments")
}

}
