% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlleleTrim.R
\name{GetField}
\alias{GetField}
\title{An HLA trimming function}
\usage{
GetField(x, Res)
}
\arguments{
\item{x}{HLA allele.}

\item{Res}{Resolution desired.}
}
\description{
Trim a properly formatted HLA allele to a desired number of fields.
}
\note{
for internal use only.
}
\examples{
GetField("HLA-A*01:01:01:01", 3)

}
