% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{getFrame}
\alias{getFrame}
\title{Access feature reading frame}
\usage{
getFrame(x, ...)
}
\arguments{
\item{x}{A \linkS4class{HLARanges} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
An \link{integer} vector.
}
\description{
Access feature reading frame
}
\examples{
##
}
