% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getStatus}
\alias{getStatus}
\title{Get feature status}
\usage{
getStatus(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{HLARanges}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector.
}
\description{
Get feature status
}
\examples{
##
}

