% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{getType}
\alias{getType}
\title{Get feature type}
\usage{
getType(x, ...)
}
\arguments{
\item{x}{A \linkS4class{HLARanges} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \link{character} vector.
}
\description{
Get feature type
}
\examples{
##
}
