% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLSC.ex}
\alias{GLSC.ex}
\title{Example Data Frame of Genotype List String Code Data}
\format{
A dataframe with 419 rows and 2 columns.
}
\source{
pould::hla.hap.demo
}
\usage{
data(GLSC.ex)
}
\description{
A two-column data frame (modified from pould::hla.hap.demo) including GL String data in Gl String Code format.
Column one identifies each subject's status, while column two identifies each subject's HLA genotype in GL String Code format.
This data is provided for use in examples and demonstrations.
}
\references{
Mack et al. HLA 2023 Oct;102(4):501-507 https://doi.org/10.1111/tan.15145
}
\keyword{datasets}
