% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLStoUNI.R
\name{GLStoUNI}
\alias{GLStoUNI}
\title{Translates strings from GL String format to UNIFORMAT format.}
\usage{
GLStoUNI(GLstring, prefix = "HLA-", pre = FALSE)
}
\arguments{
\item{GLstring}{A string of HLA allele names and operators in the GL String format, identifying their relation with one another.}

\item{prefix}{A string of the desired prefix (default is "HLA-"), this string should contain a "-" at the end.}

\item{pre}{A logical that indicates whether user would like all allele names to contain the prefix of their choice (TRUE), or if the prefix should not be appended to allele names (FALSE).}
}
\value{
An altered version of the input string, converted to UNIFORMAT format.
}
\description{
A wrapper function for GLtoUN, a function that translates strings from GL String format to UNIFORMAT format.
}
\note{
This function does not function with the "?" operator, as the "?" operator has no cognate in UNIFORMAT.
}
\examples{
GLStoUNI("HLA-A*02:01/HLA-A*02:02+HLA-A*03:01/HLA-A*03:02")
GLStoUNI("A+B/C~D^G|E^W+X/Y^Z+J")

}
\references{
Nunes Tissue Antigens 2007;69(s1):203-205 https://doi.org/10.1111/j.1399-0039.2006.00808.x

Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126
}
