% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLstring.ex}
\alias{GLstring.ex}
\title{Example Data Frame of Genotype List String Data.}
\format{
A dataframe with 419 rows and 2 columns.
}
\source{
pould::hla.hap.demo
}
\usage{
data(GLstring.ex)
}
\description{
A two-column example data frame (from pould::hla.hap.demo) including GL String Data.
Column one identifies each subject's status, while column two identifies each subject's HLA genotype in GL String format.
This data is provided for use in examples and demonstrations.
}
\references{
Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126
}
\keyword{datasets}
