% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLStoUNI.R
\name{glstringValidate}
\alias{glstringValidate}
\title{Validate a GL String.}
\usage{
glstringValidate(GLstring)
}
\arguments{
\item{GLstring}{A string of HLA allele names and operators in the GL String format signifying their relation with one another.}
}
\value{
If everything is in order, it will return the original string, if something is wrong, it will return a message and stop.
}
\description{
A function that validates that a GL String formatted string does not contain any letters or operators not supported in the GL String.
}
\examples{
glstringValidate("HLA-A*02:01/HLA-A*02:02+HLA-A*03:01/HLA-A*03:02")

}
\references{
Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126
}
