% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLupdater.R
\name{GLV2}
\alias{GLV2}
\title{Format GL String Code version number.}
\usage{
GLV2(Version)
}
\arguments{
\item{Version}{A string of the desired version to which the alleles to be updated, going back to version 1.05.0.}
}
\value{
Returns alleleListHistory$AlleleListHistory-column-formatted version.
}
\description{
Returns the AlleleListHistory-formatted version of a dot-delimited IPD-IMGT/HLA Database release, or provides appropriate options if version input is not present in the IPD-IMGT/HLA Database.
}
\note{
For internal use only.
}
\examples{
\dontrun{
GLV2("3.34.0")
GLV2("3.0.0")
}
}
\references{
Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126

Mack et al. HLA 05 July 2023 https://doi.org/10.1111/tan.15145
}
