% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLupdater.R
\name{GLvalidate}
\alias{GLvalidate}
\title{Validate a GL String Code.}
\usage{
GLvalidate(GLString)
}
\arguments{
\item{GLString}{A string formatted to GL String Code specifications.}
}
\value{
A well-formatted GL String Code, or an error message.
}
\description{
A lightweight validator to protect from incorrectly formatted GL String Codes.
}
\note{
For internal use only.
}
\examples{
GLvalidate("ha#3.25.0#hla-B15:35")

}
\references{
Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126

Mack et al. HLA 05 July 2023 https://doi.org/10.1111/tan.15145
}
