% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLupdater.R
\name{GLVhelper}
\alias{GLVhelper}
\title{Locate matches for an incomplete IPD-IMGT/HLA Database version.}
\usage{
GLVhelper(Version)
}
\arguments{
\item{Version}{A character string of the desired version to which the alleles should be updated, going back to version 1.05.0.}
}
\value{
A list of character strings of possible matches to a given incomplete input.
}
\description{
Uses the provided version information to locate and return possible matches for an incompletely defined IPD-IMGT/HLA Database release version.
}
\note{
For internal use only.
}
\examples{
GLVhelper("2.25")
GLVhelper("3.9.0")


}
\references{
Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126

Mack et al. HLA 05 July 2023 https://doi.org/10.1111/tan.15145
}
