% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.r
\name{hasNext}
\alias{hasNext}
\title{Does this iterator have a next element}
\usage{
hasNext(obj, ...)
}
\arguments{
\item{obj}{An iterator object.}

\item{...}{Additional arguments that are ignored.}
}
\value{
Logical value
}
\description{
Does this iterator have a next element
}
\examples{
##
}

