% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hla_field1}
\alias{hla_field1}
\title{Get the first field from a HLA allele code}
\usage{
hla_field1(a)
}
\arguments{
\item{alleles}{A vector of HLA allele codes (e.g.,  \emph{HLA-A*01:01:01:02}).}
}
\value{
A vector of first fields.
}
\description{
Get the first field from a HLA allele code
}

