% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hla_field2}
\alias{hla_field2}
\title{Get the second field from a HLA allele code}
\usage{
hla_field2(a)
}
\arguments{
\item{alleles}{A vector of HLA allele codes (e.g.,\emph{HLA-A*01:01:01:02}).}
}
\value{
A vector of second fields or or \emph{""} if no second field exists).
}
\description{
Get the second field from a HLA allele code
}
