% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hla_prefix}
\alias{hla_prefix}
\title{Get the HLA allele code prefix}
\usage{
hla_prefix(a)
}
\arguments{
\item{alleles}{A vector of HLA allele codes (e.g.,  \emph{HLA-A*01:01:01:02}).}
}
\value{
A vector of HLA allele code prefixes (e.g., \emph{"HLA-A"} or \emph{""} if
the allele code was not prefixed).
}
\description{
Get the HLA allele code prefix
}

