% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hla_sort}
\alias{hla_sort}
\title{Sort HLA alleles by field}
\usage{
hla_sort(alleles)
}
\arguments{
\item{alleles}{A vector of HLA alleles.}
}
\value{
A sorted vector of HLA alleles.
}
\description{
Sort HLA alleles by field
}

