% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HlaAllele.r
\docType{class}
\name{HLAAllele-class}
\alias{HLAAllele-class}
\alias{[,HLAAllele,character,missing-method}
\alias{[,HLAAllele,logical,missing-method}
\alias{[,HLAAllele,numeric,missing-method}
\alias{c,HLAAllele-method}
\title{Class \code{"HLAAllele"}}
\usage{
\S4method{c}{HLAAllele}(x, ..., recursive = FALSE)

\S4method{[}{HLAAllele,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{HLAAllele,logical,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{HLAAllele,character,missing}(x, i, j, ..., drop = TRUE)
}
\description{
Class \code{"HLAAllele"}
}
\section{Methods (by generic)}{
\itemize{
\item \code{c}: Combine \code{HLAAllele} objects.

\item \code{[}: Subset \code{HLAAllele} objects.

\item \code{[}: Subset \code{HLAAllele} objects.

\item \code{[}: Subset \code{HLAAllele} objects.
}}
\section{Slots}{

\describe{
\item{\code{sequence}}{DNAStringSet.}

\item{\code{metadata}}{DataFrame.}

\item{\code{features}}{HLARangesList.}
}}
\examples{
showClass("HLAAllele")
}
\keyword{classes}

