% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLAatlas}
\alias{HLAatlas}
\title{Boundary Positions of Exons, Introns and UTRs in Amino Acid, cDNA and gDNA Alignments}
\format{
A list of 4 elements that include the gene features boundaries in each ANHIG/IMGTHLA sequence alignment.
\itemize{
\item(prot: peptide-alignment atlases)
\item(nuc: cDNA- and codon-alignment atlases)
\item(gen: gDNA-alignment atlases)
\item(version: The IPD-IMGT/HLA Database release version under which these data were generated)
}
}
\source{
https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments
}
\usage{
data(HLAatlas)
}
\description{
A list object of two wlements. Sub-lists of R dataframes (atlases) for each locus with a protein (prot), cDNA (nuc), and gDNA (gen) alignment. Each atlas identifies the position of the exon (E), intron (I) or UTR (U) gene-feature boundaries in an alignment. Boundaries for non-standard hybrid (H), join (J), novel (N) and segment (S) features may be included in gene fragment and pseudogene atlases.
This object is built by the atlasFull() function.
}
\keyword{datasets}
