% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLAgazeteer}
\alias{HLAgazeteer}
\title{Identifies HLA loci that have amino acid, cDNA, or gDNA alignments.}
\format{
A list object containing 9 character vectors
}
\source{
https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments

HLAtools::IMGTHLAGeneTypes
}
\usage{
data(HLAgazeteer)
}
\description{
A list object of vectors that identify loci with amino acid (HLAgazeteer$prot), cDNA (HLAgazeteer$nuc), gDNA alignments (HLAgazeteer$gen), pseudogenes (HLAgazeteer$pseudo), gene fragments (HLAgazeteer$frag), as well as pseudogenes and gene fragments for which no cDNA ($noNuc) and gDNA ($noGen) alignments are available).
This object is built by the buildGazeteer() function in the HLAtools package.
}
\note{
Pseudogenes and gene fragments derived from the IMGTHLAGeneTypes object.
}
\keyword{datasets}
