% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAGene.R
\name{HLAGene}
\alias{HLAGene}
\title{Constructor for \link[=HLAGene_]{HLAGene} objects.}
\usage{
HLAGene(
  locusname,
  db_version = "Latest",
  db_path = getOption("hlatools.local_repos"),
  ...
)
}
\arguments{
\item{locusname}{<\link{character}>; a valid HLA gene name.}

\item{db_version}{<\link{character}>; IPD-IMGT/HLA version (e.g.: 3.28.0, 3.18.0, ...)}

\item{db_path}{<\link{character}>; Location of local IPD-IMGT/HLA repository. Defaults to \file{~/local/db/IMGTHLA}.}

\item{...}{Passed on to the initialiser.}
}
\value{
A \link[=HLAGene_]{HLAGene} object.
}
\description{
Constructor for \link[=HLAGene_]{HLAGene} objects.
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
db_version(x)
locusname(x)
length(x)

## extract alleles with complete sequences
x[is_complete(x)]

## extract all Common alleles
x[cwd_status(x) == "Common"]
}
}
