% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAGene.r
\docType{class}
\name{HLAGene_}
\alias{HLAGene_}
\title{Class \code{"HLAGene"}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HLAgene(locusname)
}
\value{
Object of \code{\link{R6Class}} representing an HLA gene.
}
\description{
Class \code{"HLAGene"}
}
\section{Fields}{

\describe{
\item{\code{locusname}}{\code{[character]}; A valid HLA gene.}

\item{\code{alleles}}{\code{[HLAAllele]}; An \code{\linkS4class{HLAAllele}} object.}

\item{\code{dm}}{\code{[matrix]}; Genetic distances between exons 2.}

\item{\code{cons}}{\code{[DNAStringSet]}; Strict consensus of completely sequenced alleles.}
}}
\section{Methods}{

\describe{
  \item{\code{x$new(locusname)}}{Create object of this class.}
  \item{\code{x$get_allele(allele)}}{Extract an allele.}
  \item{\code{x$get_closest_complete_neighbor(allele)}}{Get the complete allele that closest at exon 2 to the query allele}
  \item{\code{x$get_reference_sequence(allele)}}{}
}
}
\keyword{data}
\keyword{internal}

