% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLARanges.r
\name{HLARanges}
\alias{HLARanges}
\title{Constructor for \code{\linkS4class{HLARanges}} objects}
\usage{
HLARanges(seqnames = Rle(), ranges = IRanges(), id = NA_character_,
  order = NA_integer_, type = NA_character_, status = NA_character_,
  frame = NA_integer_, ...)
}
\arguments{
\item{seqnames}{character vector of the sequence name.}

\item{ranges}{IRanges object containing the ranges.}

\item{id}{character vector of feature IDs.}

\item{order}{integer vector of feature order.}

\item{type}{character vector of feature type.}

\item{status}{character vector of completeness status}

\item{frame}{integer vector of reading order.}

\item{...}{Arguments passed to \code{\link[GenomicRanges]{Granges}}.}
}
\value{
A \code{\linkS4class{HLARanges}} object.
}
\description{
Constructor for \code{\linkS4class{HLARanges}} objects
}
\examples{
showClass("HLARanges")
}
\seealso{
\code{\link{parse_hla_alleles}}, \code{\link{HLAAllele}}
}

