% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLARanges.r
\docType{class}
\name{HLARangesList-class}
\alias{HLARangesList-class}
\title{Class \code{"HLARangesList"}}
\description{
Class \code{"HLARangesList"}
}
\examples{
showClass("HLARangesList")
}
\keyword{classes}

