% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLARanges.r
\name{HLARangesList}
\alias{HLARangesList}
\title{Construct HLARangesList objects}
\usage{
HLARangesList(...)
}
\arguments{
\item{...}{HLARanges objects.}
}
\value{
A HLARangesList object
}
\description{
Construct HLARangesList objects
}
\examples{
###
}

