% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlatools-package.r
\docType{package}
\name{hlatools-package}
\alias{hlatools}
\alias{hlatools-package}
\title{Collection of tools to work with HLA data}
\description{
Collection of tools to work with HLA data.
}
\details{
Details
}

