% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{hlatools_version}
\alias{hlatools_version}
\title{Get hlatools package version}
\usage{
hlatools_version(x, ...)
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \link{package_version} object.
}
\description{
Access the package version under which a \link[=HLAGene_]{HLAGene} object has been
created.
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
hlatools_version(x)
}
}
