% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HTexceptions}
\alias{HTexceptions}
\title{Defines specific exceptions for use with package functions}
\format{
A list of elements containing vectors of data for specific exceptions.
\itemize{
\item(pseudo.codon: A vector of pseudogenes that include spurious codon alignment coordinates in some releases)
\item(date: A YYYY-MM-DD_HH:mm:ss.SSSSSS formatted date stamp describing the HTexceptions object's time of creation)
}
}
\source{
HLAtools:HTexceptions()
}
\usage{
data(HTexceptions)
}
\description{
A list object of sub-lists of vectors that define exceptions for specific use cases.
}
\keyword{datasets}
