% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{ihasNext}
\alias{ihasNext}
\title{Create an iterator with a hasNext method}
\usage{
ihasNext(it)
}
\arguments{
\item{it}{An iterable object.}
}
\value{
An \code{ihasNext} iterator
}
\description{
Create an iterator with a hasNext method
}
\examples{
###
}
