% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMGTGenomicReferences}
\alias{IMGTGenomicReferences}
\title{Genomic Reference Alleles for each Gene in the IPD-IMGT/HLA Database}
\format{
A list of two elements.
\itemize{
\item(GenomicReferenceGenes: A table of 44 rows, one for each gene, and 6 columns, identifying each gene, its Ensembl version, GRChg38 location, genomic coordinate of that gene's start codon, GRChg38 reference allele, and IPD-IMGT/HLA reference allele)
\item(Version: The date on which this object was generated)
}
}
\source{
https://www.ebi.ac.uk/ipd/imgt/hla/about/help/genomics/
}
\usage{
data(IMGTGenomicReferences)
}
\description{
A data.frame characterizing a genomic reference allele for each IPD-IMGT/HLA Database gene, and a vector identifying the date that the table was generated.
This object is built by the buildGenomicReferences() function.
}
\keyword{datasets}
