% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMGTHLAGeneTypes}
\alias{IMGTHLAGeneTypes}
\title{Describes the molecular characteristics of the genes curated by the IPD-IMGT/HLA Database, and identifies gene fragments and pseudogenes.}
\format{
A data frame of three columns
}
\source{
https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments
}
\usage{
data(IMGTHLAGeneTypes)
}
\description{
A data frame of three columns identifying each gene supported by the IPD-IMGT/HLA Database, its molecular characteristics, and its status as gene fragment or pseudogene.
This object is built by the BuildIMGTHLAGeneTypes() function.
}
\keyword{datasets}
