% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{intron}
\alias{intron}
\title{Retrieve intron sequences}
\usage{
intron(x, intron = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{HLAGene}} object.}

\item{intron}{The feature number or \code{NULL} (Retrieves all introns).}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\linkS4class{DNAStringSet}}.
}
\description{
Retrieve intron sequences
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
intron(x, 1:2)
}
}
