% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{is_complete}
\alias{is_complete}
\title{Access completeness status}
\usage{
is_complete(x, ...)
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} or \linkS4class{HLAAllele} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A logical vector.
}
\description{
Access completeness status
}
\note{
A complete allelic sequence is defined as a sequence that includes
3'-UTR and 5'-UTR sequence, irrespective of whether the UTR sequences
are complete.
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
x[is_complete(x)]
}
}
