% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{is_lsl}
\alias{is_lsl}
\title{Has a sequence been submitted by DKMS Life Science Lab}
\usage{
is_lsl(x, ...)
}
\arguments{
\item{x}{A \code{\link{HLAGene}} or \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A logical vector.
}
\description{
Has a sequence been submitted by DKMS Life Science Lab
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
x[is_lsl(x) & is_complete(x)]
}
}
