% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{locusname}
\alias{locusname}
\alias{locusname<-}
\title{Access locus name}
\usage{
locusname(x, ...)

locusname(x, ...) <- value
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} object.}

\item{...}{Further arguments passed to methods.}

\item{value}{A valid HLA locus name or \code{NA}.}
}
\value{
A character vector.
}
\description{
Access locus name
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
locusname(x)
}
}
