% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_composite.R
\name{make_tree_input}
\alias{make_tree_input}
\title{Convert genotype predictions into tree input format
Essentially one-hot encoding of the genotype column}
\usage{
make_tree_input(hla)
}
\arguments{
\item{hla}{A data frame with columns: \cr
"sample" - unique sample name \cr
"locus" - one of the following HLA loci: "A", "B", "C", "DPA1", "DPB1", "DQA1", "DQB1", "DRB1", "DRB3", "DRB4", "DRB5" \cr
"field" - the resolution of the hla allele in the format of "field_1", "field_2", "field_3" \cr
"genotyper" - one of the following genotyping algorithms "arcasHLA", "hlaminer", "optitype", "phlat" \cr
"allele" - a character representing the predicted allele with fields separated by "_" (e.g. "01", "01_02", "01_02_01") \cr}
}
\value{
A data frame with columns: \cr
"sample" - unique sample name \cr
"locus" - one of the following HLA loci: "A", "B", "C", "DPA1", "DPB1", "DQA1", "DQB1", "DRB1", "DRB3", "DRB4", "DRB5" \cr
"field" - the resolution of the hla allele in the format of "field_1", "field_2", "field_3" \cr
"present_arcasHLA", "present_hlaminer", "present_optitype", "present_phlat" - a T/F value indicating if a prediction was present for the indicated genotyping algorithm
}
\description{
Convert genotype predictions into tree input format
Essentially one-hot encoding of the genotype column
}
\examples{
data(composite_data)
hlatools:::make_tree_input(composite_data)
}
