% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_alleles}
\alias{match_alleles}
\title{(Partially) match allele in a HLAGene object.}
\usage{
match_alleles(allele, x, partially = FALSE)
}
\arguments{
\item{allele}{\link{character}; A vector of (partial) allele names.}

\item{x}{A \linkS4class{HLAAllele} or \linkS4class{HLAGene} object.}

\item{partially}{If \code{TRUE} match partial allele names.}
}
\value{
An \link{integer} vector
}
\description{
(Partially) match allele in a HLAGene object.
}
\examples{
##
}
\keyword{internal}
