% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDNASequenceSearch.R
\name{multicDNAalign}
\alias{multicDNAalign}
\title{Generate a codon alignment for specific HLA alleles at different sets of positions.}
\usage{
multicDNAalign(alleles, positions)
}
\arguments{
\item{alleles}{A vector of un-prefixed HLA locus names.}

\item{positions}{A list of vectors of nucleotide positions, exactly one vector for each allele, against which each corresponding allele will be aligned.}
}
\value{
A data frame of 'allele' and the corresponding nucleotide sequence for specified positions designated for an allele. a nucleotide sequence is not returned for a specific allele if input allele is not available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a nucleotide at the input position.
}
\description{
Generates a codon (cDNA) alignment for HLA alleles allowing each allele to be aligned to a different set of positions.
}
\note{
For internal HLAtools use only.
}
\examples{
\dontrun{
multicDNAalign(c("DQA1*01:01:01:01", "DRB1*01:01:01:01"),list(32:58, 33:59))
}
}
