% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDNASequenceSearch.R,
%   R/cDNASequenceSearch_2_SM.R
\name{multicDNAsearch}
\alias{multicDNAsearch}
\title{Search HLA codon sequences at multiple positions for a specified HLA allele}
\usage{
multicDNAsearch(
  alignType,
  locus,
  allele,
  positions,
  prefix = TRUE,
  sep = "~",
  trimmed = FALSE
)

multicDNAsearch(
  alignType,
  locus,
  allele,
  positions,
  prefix = TRUE,
  sep = "~",
  trimmed = FALSE
)
}
\arguments{
\item{alignType}{The type of alignment being searched. Allowed values are "codon" and "nuc".}

\item{locus}{A specific HLA locus.}

\item{allele}{The name of the allele to search for.}

\item{positions}{Nucleotide positions to search.}

\item{prefix}{A logical that indicates if the position number(s) should be included in the result.}

\item{sep}{Defines the separator between position sequences.}

\item{trimmed}{A logical identifying whether 'allele' is a two-field, 'trimmed' allele name (trimmed=TRUE), or a full-length name (trimmed=FALSE). The default value is FALSE.}
}
\value{
codons (possibly with position number) at 'positions' for 'allele'. If 'allele' does not have a codon at 'position', '' will be returned.

codons (possibly with position number) at 'positions' for 'allele'. If 'allele' does not have a codon at 'position', '' will be returned.
}
\description{
Generates a character string of codon or nuc (cDNA) sequences at the specified positions for the specified HLA allele name.

Generates a character string of codon or nuc (cDNA) sequences at the specified positions for the specified HLA allele name.
}
\note{
For internal HLAtools use only.

For internal HLAtools use only.
}
\examples{
\dontrun{
multicDNAsearch("codon", "DRB1", "15:07",c(23:26), trimmed=TRUE)
multicDNAsearch("codon", "DRB1", "15:07:01",c(23:26))
}
\dontrun{
multicDNAsearch("codon", "DRB1", "15:07",c(23:26), trimmed=TRUE)
multicDNAsearch("codon", "DRB1", "15:07:01",c(23:26))
}
}
