% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDNASequenceSearch.R
\name{multigDNAalign}
\alias{multigDNAalign}
\title{Generate a genomic alignment for specific HLA alleles at different sets of positions.}
\usage{
multigDNAalign(alleles, positions)
}
\arguments{
\item{alleles}{A vector of un-prefixed HLA allele names.}

\item{positions}{A list of vectors of nucleotide positions, exactly one vector for each allele, against which each corresponding allele will be aligned.}
}
\value{
A data frame of 'allele' and the corresponding nucleotide sequence for 'positions'. a nucleotide sequence is not returned for a specific allele if input allele is not available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a nucleotide at the input position.
}
\description{
Generates a genomic (gDNA) nucleotide alignment for HLA alleles allowing each allele to be aligned to a different set of positions.
}
\note{
for internal use only
}
\examples{
\dontrun{
multigDNAalign(c("DQA1*01:01:01:01", "DRB1*01:01:01:01"),list(c(25,46,50,78,88), c(26,47,51,79,89)))
multigDNAalign(c("DQA1*01:01:01:01", "DRB1*01:01:01:01"),list(32:58, 33:59))
multigDNAalign(c("DPB1*01:01:01:01","DQB1*05:01:01:01","DRB1*01:01:01:01"),list(32:58,32:58,32:58))
}
}
