% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDNASequenceSearch.R
\name{multigDNAsearch}
\alias{multigDNAsearch}
\title{Search genomic HLA nucleotide sequences at multiple positions for a specified HLA allele}
\usage{
multigDNAsearch(
  locus,
  allele,
  positions,
  prefix = TRUE,
  sep = "~",
  trimmed = FALSE
)
}
\arguments{
\item{locus}{An HLA locus in the HLAalignments$gen object in the HLAtools.data package.}

\item{allele}{The name of an allele at 'locus'.}

\item{positions}{A set of consecutive (e.g., 1:10) or non consecutive (e.g., c(1,3,34,344)) nucleotide positions in the 'locus' alignment.}

\item{prefix}{A logical that indicates if the position number(s) should be included in the result.}

\item{sep}{Defines the separator between position sequences. The default separator is '~'.}

\item{trimmed}{A logical identifying whether 'allele' is a two-field, 'trimmed' allele name (trimmed=TRUE), or a full-length name (trimmed=FALSE). The default value is FALSE.}
}
\value{
nucleotides (possibly with position number) at 'positions' for 'allele'. If 'allele' does not have a nucleotide at 'position', '' will be returned.
}
\description{
Generates a character string of genomic (gDNA) nucleotide sequences at the specified positions for the specified HLA allele name.
}
\note{
For internal HLAtools use only.
}
\examples{
\dontrun{
multigDNAsearch("DRB1", "15:07",c(23:26), trimmed=TRUE)
multigDNAsearch("DRB1", "15:07:01",c(23:26))
}
}
