% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLStoUNI.R
\name{multiGLStoUNI}
\alias{multiGLStoUNI}
\title{Translates columns of GL Strings to UNIFORMAT strings.}
\usage{
multiGLStoUNI(GLstringArray, prefix = "HLA-", pre = FALSE)
}
\arguments{
\item{GLstringArray}{A vector with GL String format. If the vector has more than one column, the first column should only contain labels/identifiers; the first column of a multi-column vector should NOT contain GL Strings. If the input vector only contains one column, the column should contain GL Strings.}

\item{prefix}{A string of the desired prefix (autoset to "HLA-"), this string should contain a "-" at the end.}

\item{pre}{A logical that indicates whether user would like all allele names to contain the prefix of their choice (TRUE), or if the prefix should not be appended to allele names (FALSE).}
}
\value{
An altered version of the input array, converted to UNIFORMAT format.
}
\description{
A function that translates columns of arrays containing strings in GL String format to UNIFORMAT format.
}
\note{
This function does not function with the "?" operator, as the "?" operator has no cognate in UNIFORMAT.
}
\examples{
multiGLStoUNI(GLstring.ex[[2]][1:5])

}
\references{
Nunes Tissue Antigens 2007;69(s1):203-205 https://doi.org/10.1111/j.1399-0039.2006.00808.x

Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126
}
