% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AminoAcidSequenceSearch.R
\name{multipleAAalign}
\alias{multipleAAalign}
\title{Generate an amino acid alignment for specific HLA alleles at different sets of positions.}
\usage{
multipleAAalign(alleles, positions)
}
\arguments{
\item{alleles}{A vector of un-prefixed HLA locus names.}

\item{positions}{A list of vectors of AA positions, exactly one vector for each allele, against which each corresponding allele will be aligned.}
}
\value{
A data frame of allele names and the corresponding peptide sequence for each amino acid position for the specified positions designated for an allele. a peptide sequence is not returned for a specific allele if input allele is not available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a peptide at the input position.
}
\description{
Generates an amino acid alignment for HLA alleles allowing each allele to be aligned to a different set of positions.
}
\note{
For internal HLAtools use only.
}
\examples{
\dontrun{
multipleAAalign(c("DQA1*01:01","DQB1*05:01","DPB1*01:01"),list(32:58,33:59,31:57))
}
}
