% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AminoAcidSequenceSearch.R
\name{multipleAAsearch}
\alias{multipleAAsearch}
\title{Search HLA amino acid sequences at multiple positions for a specified HLA allele}
\usage{
multipleAAsearch(
  locus,
  allele,
  positions,
  prefix = TRUE,
  sep = "~",
  trimmed = FALSE
)
}
\arguments{
\item{locus}{Specific locus of allele.}

\item{allele}{Name of allele to search for.}

\item{positions}{Amino-acid positions to search.}

\item{prefix}{A logical that indicates if the position number(s) should be included in the result.}

\item{sep}{Defines the separator between position sequences}

\item{trimmed}{A logical identifying whether 'allele' is a two-field, "trimmed" allele name (trimmed=TRUE), or a full-length name (trimmed=FALSE). The default value is FALSE.}
}
\value{
Amino-acids (possibly with position number) at specified positions for specified allele as available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a peptide at the input position.
}
\description{
Generates a character string of amino acid (AA) sequences at the specified positions for the specified HLA allele name.
}
\note{
For internal HLAtools use only.
}
\examples{
\dontrun{
multipleAAsearch("DRB1", "15:07:01", 11:22)
}
}
