% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNItoGLS.R
\name{multiUNItoGLS}
\alias{multiUNItoGLS}
\title{Translate columns of UNIFORMAT strings to GL Strings.}
\usage{
multiUNItoGLS(UniformatArray, prefix = "HLA-", pre = TRUE)
}
\arguments{
\item{UniformatArray}{A vector with UNIFORMAT formatted strings. if the vector has more than one column, the first column should only contain labels/identification data. The first column of a multi-column vector should NOT contain UNIFORMAT strings. if the input vector only contains one column, the column should contain UNIFORMAT strings.}

\item{prefix}{A string of the desired prefix (autoset to "HLA-"), this string should contain a "-" at the end.}

\item{pre}{A logical that indicates whether user would like all allele names to contain the prefix of their choice (TRUE), or if the prefix should not be appended to allele names (FALSE).}
}
\value{
An altered version of the input string, converted to GL String format.
}
\description{
A function that translates columns of arrays containing UNIFORMAT formatted strings to GL String format.
}
\note{
This function does not return the "?" operator, as the "?" operator has no cognate in UNIFORMAT.
}
\examples{
multiUNItoGLS(UNIFORMAT.example[1:2])
multiUNItoGLS(UNIFORMAT.example[2])

}
\references{
Nunes Tissue Antigens 2007;69(s1):203-205 https://doi.org/10.1111/j.1399-0039.2006.00808.x

Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126
}
