% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSSample.r
\name{NGSSample}
\alias{NGSSample}
\title{Constructor for \code{\link[=NGSSample_]{NGSSample}} objects}
\usage{
NGSSample(xmlpath, lims_donor_id = NULL)
}
\arguments{
\item{xmlpath}{Path to an NGSengine NGSSample XML file.}

\item{lims_donor_id}{[Optional] Sample ID.}
}
\value{
A \code{\link[=NGSSample_]{NGSSample}} object.
}
\description{
Constructor for \code{\link[=NGSSample_]{NGSSample}} objects
}
\examples{
\dontrun{
NGSSample("path-to-NGSengine-xml")
}
}

