% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSSample.r
\docType{class}
\name{NGSSample_}
\alias{NGSSample_}
\title{Class \code{"NGSSample"}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NGSSample(xmlpath)
}
\value{
Object of \code{\link{R6Class}} representing an NGSengine NGSSample XML.
}
\description{
Class \code{"NGSSample"}
}
\section{Fields}{

\describe{
\item{\code{path}}{\code{[character]}.}

\item{\code{metadata}}{\code{[list]}.}

\item{\code{region}}{\code{[Biostrings::BStringSet]}.}

\item{\code{genotypes}}{\code{[list]}.}
}}
\section{Methods}{

\describe{
  \item{\code{x$new(xmlpath)}}{Create object of this class.}
  \item{\code{x$allele1()}}{First allele.}
  \item{\code{x$allele2()}}{Second allele.}
  \item{\code{x$locusname()}}{}
  \item{\code{x$samplename()}}{}
  \item{\code{x$lims_donor_id()}}{}
  \item{\code{x$xml_path()}}{}
}
}
\keyword{data}
\keyword{internal}

