% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{noutr}
\alias{noutr}
\title{Access all sequence data except for the UTRs}
\usage{
noutr(x, ...)
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{DNAStringSet}.
}
\description{
Access all sequence data except for the UTRs
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
noutr(x)
}
}
