% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{p_group}
\alias{p_group}
\title{Access P group designations}
\usage{
p_group(x, ...)
}
\arguments{
\item{x}{A \code{\link{HLAGene}} or \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of allele names.
}
\description{
Access P group designations
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
p_group(x[is_complete(x)])
}
}

