% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hla_alleles.r
\name{parse_hla_alleles}
\alias{parse_hla_alleles}
\title{Parse all HLA alleles for a locus from hla.xml}
\usage{
parse_hla_alleles(doc, locusname, ncores = parallel::detectCores())
}
\arguments{
\item{doc}{\file{hla.xml} as an \code{xml_document} object.}

\item{locusname}{One of [\code{HLA-A, HLA-B, HLA-C, HLA-DPA1, HLA-DPB1,
HLA-DQA1, HLA-DQB1, HLA-DRB}]}

\item{ncores}{he number of compute cores to use.}
}
\value{
A \code{\linkS4class{HLAAllele}} object.
}
\description{
Parse all HLA alleles for a locus from hla.xml
}
\examples{
\dontrun{
doc <- read_hla_xml()
dpb1 <- parse_hla_alleles(doc, "HLA-DPB1")
}
}

