% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryFunctions.R
\name{posSort}
\alias{posSort}
\title{Numerical sort for sequence alignment positions that contain indels.}
\usage{
posSort(posVec, alignType, locus)
}
\arguments{
\item{posVec}{A vector of nucleotide positions}

\item{alignType}{The type of alignment the positions are in. The values 'prot', 'nuc' and 'gen' are valid options.}

\item{locus}{A locus in the HLAalignments object for the specified alignType.}
}
\value{
A correctly sorted sequence.
}
\description{
Sorts sequence alignment positions that contain indels in numerical order; e.g., three indels following position X are are identified as X.1, X.2 and X.3.
}
\examples{
\dontrun{
posSort(c(2,4,3,1,5), "nuc","DRB1")
}
}
