% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_composite.R
\name{predict_composite}
\alias{predict_composite}
\title{Predict the best genotyping algorithm for each allele prediction}
\usage{
predict_composite(hla)
}
\arguments{
\item{hla}{A data frame with columns: \cr
"sample" - unique sample name \cr
"locus" - one of the following HLA loci: "A", "B", "C", "DPA1", "DPB1", "DQA1", "DQB1", "DRB1", "DRB3", "DRB4", "DRB5" \cr
"field" - the resolution of the hla allele in the format of "field_1", "field_2", "field_3" \cr
"genotyper" - one of the following genotyping algorithms "arcasHLA", "hlaminer", "optitype", "phlat" \cr
"allele" - a character representing the predicted allele with fields separated by "_" (e.g. "01", "01_02", "01_02_01") \cr}
}
\value{
A data frame with columns: \cr
"sample" - unique sample name \cr
"locus" - one of the following HLA loci: "A", "B", "C", "DPA1", "DPB1", "DQA1", "DQB1", "DRB1", "DRB3", "DRB4", "DRB5" \cr
"field" - the resolution of the hla allele in the format of "field_1", "field_2", "field_3" \cr
"best_genotyper" - the best genotyping out of "arcasHLA", "hlaminer", "optitype", "phlat" as determined by the tree model
}
\description{
Predict the best genotyping algorithm for each allele prediction
}
\examples{
data(composite_data)
predict_composite(composite_data)
}
