% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_drb345.R
\name{predict_drb345}
\alias{predict_drb345}
\title{Predict DRB345 copy numbers}
\usage{
predict_drb345(drb)
}
\arguments{
\item{drb}{A data frame with columns: \cr
"sample" - unique sample name \cr
"DRB1", "DRB3", "DRB4", "DRB5" - raw read counts for each loci}
}
\value{
A data frame with columns: \cr
"sample" - unique sample name \cr
"DRB3", "DRB4", "DRB5" - copy number of each loci represented as value of 0,1,2
}
\description{
Predict DRB345 copy numbers
}
\examples{
data(drb345_data)
predict_drb345(drb345_data)

}
