% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDtoPyPop.R
\name{pypopHeaders}
\alias{pypopHeaders}
\title{Convert the header of a BIGDAWG data file to PyPop format.}
\usage{
pypopHeaders(colHead)
}
\arguments{
\item{colHead}{A vector of column names. BIGDAWG format requires that the first two header values are the sample identifier character string and the subject status (0 or 1). All other header values should be paired locus/gene names.}
}
\value{
A PyPop-formatted column header.
}
\description{
Convert the header of a BIGDAWG-formatted data frame into PyPop-formatted header.
}
\examples{
\dontrun{
pypopHeaders(colnames(BIGDAWG::HLA_data))
}
}
\references{
Lancaster et al. Front Immunol. 2024 Apr 2;15:1378512. https://pubmed.ncbi.nlm.nih.gov/38629078/

Pappas et al. Hum Immunol. 2016 Mar 77(3):283-287. https://pubmed.ncbi.nlm.nih.gov/26708359/
}
