% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HLA_functions.R
\name{read_arcasHLA}
\alias{read_arcasHLA}
\title{Read genotype output from arcasHLA}
\usage{
read_arcasHLA(dir)
}
\arguments{
\item{dir}{Path to arcasHLA output directory}
}
\value{
A data frame
}
\description{
Read genotype output from arcasHLA
}
\examples{
# No example
}
