% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_drb345.R
\name{read_arcasHLA_reads}
\alias{read_arcasHLA_reads}
\title{Read HLA locus alignment statistics from arcasHLA}
\usage{
read_arcasHLA_reads(dir)
}
\arguments{
\item{dir}{Path to arcasHLA output directory}
}
\value{
data frame
}
\description{
Read HLA locus alignment statistics from arcasHLA
}
\examples{
# No example
}
