% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.r
\name{read_hla_xml}
\alias{read_hla_xml}
\title{Parse the IMGT/HLA hla.xml file}
\usage{
read_hla_xml(remote = FALSE)
}
\arguments{
\item{remote}{[logical] Pull data from the IMGT/HLA ftp server or
\code{getOption("hlatools.local_repos")}}
}
\value{
An object of class (S3) \code{XMLInternalDocument}.
}
\description{
Parse the IMGT/HLA hla.xml file
}
\examples{
doc <- read_hla_xml(remote = TRUE)
}

