% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HLA_functions.R
\name{read_HLAminer}
\alias{read_HLAminer}
\title{Read genotype output from PHLAT}
\usage{
read_HLAminer(dir)
}
\arguments{
\item{dir}{File path to a single HLAminer output file}
}
\value{
data frame
}
\description{
Read genotype output from PHLAT
}
\examples{
# No examples
}
