% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HLA_functions.R
\name{read_HLAminer_single_sample}
\alias{read_HLAminer_single_sample}
\title{Read HLAminer output for a single file
Separated single file read into a single function since parsing
HLAminer output is complicated and would be difficult to read inside
of the function to parse the entire directory.
Meant to be used internally.}
\usage{
read_HLAminer_single_sample(path)
}
\arguments{
\item{path}{Path to HLAminer output directory}
}
\value{
data frame
}
\description{
Read HLAminer output for a single file
Separated single file read into a single function since parsing
HLAminer output is complicated and would be difficult to read inside
of the function to parse the entire directory.
Meant to be used internally.
}
\examples{
# No example
}
