% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HLA_functions.R
\name{read_OptiType}
\alias{read_OptiType}
\title{Read genotype output from OptiType}
\usage{
read_OptiType(dir)
}
\arguments{
\item{dir}{Path to OptiType output directory}
}
\value{
data frame
}
\description{
Read genotype output from OptiType
}
