% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HLA_functions.R
\name{read_PHLAT}
\alias{read_PHLAT}
\title{Read genotype output from PHLAT}
\usage{
read_PHLAT(dir)
}
\arguments{
\item{dir}{Path to PHLAT output directory}
}
\value{
data frame
}
\description{
Read genotype output from PHLAT
}
\examples{
# No example
}
