% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLARanges.r
\docType{import}
\name{reexports}
\alias{elementMetadata}
\alias{elementMetadata<-}
\alias{end}
\alias{ranges}
\alias{ranges<-}
\alias{reexports}
\alias{seqnames}
\alias{start}
\alias{width}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GenomicRanges}{\code{\link[GenomicRanges]{start}}, \code{\link[GenomicRanges]{end}}, \code{\link[GenomicRanges]{width}}, \code{\link[GenomicRanges]{seqnames}}, \code{\link[GenomicRanges]{ranges}}, \code{\link[GenomicRanges]{ranges<-}}, \code{\link[GenomicRanges]{elementMetadata}}, \code{\link[GenomicRanges]{elementMetadata<-}}}
}}
\keyword{internal}

