% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{sample_name}
\alias{sample_name}
\title{Access (colon-separated) sample names}
\usage{
sample_name(x, ...)
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} or \linkS4class{HLAAllele} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of colon-separated sample ethnicities.
}
\description{
Access (colon-separated) sample names
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
sample_name(x[is_complete(x)])
}
}
