% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{sequences}
\alias{sequences}
\alias{sequences<-}
\title{Access sequence data}
\usage{
sequences(x, ...)

sequences(x, ...) <- value
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} or \linkS4class{HLAAllele} object.}

\item{...}{Further arguments passed to methods.}

\item{value}{A \linkS4class{DNAStringSet} object.}
}
\value{
A \linkS4class{DNAStringSet} object.
}
\description{
Access sequence data
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
sequences(x[is_complete(x)])
}
}
