% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sHLAdata}
\alias{sHLAdata}
\title{Synthetic HLA Data for use with Package Examples}
\format{
A data frame of 18 columns and 47 rows.
}
\source{
These synthetic data were generated as part of the 13th International HLA Workshop to demonstrate PyPop functions. See \url{http://pypop.org/docs/guide-chapter-usage.html#data-minimal-noheader} for additional details.
}
\usage{
data(sHLAdata)
}
\description{
A BIGDAWG-formatted data frame of 18 columns and 47 rows containing synthetic HLA-A, -C, -B, -DRB1, -DQA1, -DQB1, -DPA1, and -DPB1 genotype data for 24 control subjects and 23 case subjects. Allele name data were recorded under IPD-IMGT/HLA Database release version 3.56.0. These are synthetic data generated for the imaginary UchiTelle population, and do not represent biologically-derived HLA genotypes.
}
\keyword{datasets}
