% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AminoAcidSequenceSearch.R
\name{singleAAalign}
\alias{singleAAalign}
\title{Generate an amino acid alignment for specific alleles at specific positions.}
\usage{
singleAAalign(alleles, positions)
}
\arguments{
\item{alleles}{A vector of un-prefixed HLA locus names.}

\item{positions}{A vector of amino acid positions, against which all loci will be aligned.}
}
\value{
A data frame of allele names and the corresponding peptide sequence for each amino acid position. a peptide sequence is not returned for a specific allele if input allele is not available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a peptide at the input position.
}
\description{
Generates an amino acid alignment at a single set of amino acid positions for HLA alleles at one or more loci.
}
\note{
For internal HLAtools use only.
}
\examples{
\dontrun{
singleAAalign(c("DRB1*01:01","DQB1*02:01","DPB1*01:01"),c(32:58))
}
}
