% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AminoAcidSequenceSearch.R
\name{singleAAsearch}
\alias{singleAAsearch}
\title{Search HLA amino acid sequences at single specified position for a specified HLA allele}
\usage{
singleAAsearch(locus, allele, position, prefix = TRUE, trimmed = FALSE)
}
\arguments{
\item{locus}{Specific locus of allele.}

\item{allele}{Name of allele to search for.}

\item{position}{Amino-acid position.}

\item{prefix}{A logical that indicates if the position number(s) should be included in the result.}

\item{trimmed}{A logical identifying whether 'allele' is a two-field, "trimmed" allele name (trimmed=TRUE), or a full-length name (trimmed=FALSE). The default value is FALSE.}
}
\value{
Amino-acid (possibly with position number) at specified position for specified allele as available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a peptide at the input position.
}
\description{
Generates a character string of the amino acid (AA)  sequence at the specified position for the specified HLA allele name.
}
\note{
For internal HLAtools use only.
}
\examples{
\dontrun{
singleAAsearch("DRB1", "15:07:01", 11)
}
}
