% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionValidation.R
\name{squashVersion}
\alias{squashVersion}
\title{Reduce a Release Version to Numerals}
\usage{
squashVersion(ver, num = FALSE)
}
\arguments{
\item{ver}{A character vector-formatted IPD-IMGT/HLA Database Release Version name (e.g., '3.54.0').}

\item{num}{A Logial. A numeric value is returned when num = TRUE. A character vector is returned when num = FALSE. The default value is FALSE.}
}
\value{
A 4-digit value, as either a character string or a number.
}
\description{
Removes the '.' delimiters from an IPD-IMGT/HLA Database Release Version name
}
\note{
For internal HLAtools use.
}
\examples{
squashVersion("3.45.0",TRUE)

}
