% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDNASequenceSearch.R
\name{unicDNAsearch}
\alias{unicDNAsearch}
\title{Search HLA codon sequences at single specified position for a specified HLA allele}
\usage{
unicDNAsearch(locus, allele, position, prefix = TRUE, trimmed = FALSE)
}
\arguments{
\item{locus}{A specific HLA locus.}

\item{allele}{The name of the allele to search for.}

\item{position}{Nucleotide position.}

\item{prefix}{A logical that indicates if the position number(s) should be included in the result.}

\item{trimmed}{A logical identifying whether 'allele' is a two-field, 'trimmed' allele name (trimmed=TRUE), or a full-length name (trimmed=FALSE). The default value is FALSE.}
}
\value{
Three nucleotides (possibly with position number) at specified position for specified allele as available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a nucleotide at the input position.
}
\description{
Generates a character string of the codon (cDNA)  sequence at the specified position for the specified HLA allele name.
}
\note{
For internal HLAtools use only.
}
\examples{
\dontrun{
unicDNAsearch("DRB1", "15:07", 57, trimmed=TRUE)
unicDNAsearch("DRB1", "15:07:01", 57)
}
}
