% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNIFORMAT.example}
\alias{UNIFORMAT.example}
\title{Example Data Frame of UNIFORMAT Data.}
\format{
A dataframe with 20 rows and 2 columns.
}
\source{
https://hla-net.eu/wp/wp-content/uploads/example-three-loci.unif_.txt
}
\usage{
data(UNIFORMAT.example)
}
\description{
A two-column example data frame including UNIFORMAT data.
Column one contains sample identifiers, while column two identifies each subject's HLA genotype in UNIFORMAT format.
To be used for example runs and demonstrations.
}
\references{
Nunes Tissue Antigens 2007;69(s1):203-205 https://doi.org/10.1111/j.1399-0039.2006.00808.x
}
\keyword{datasets}
