% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNItoGLS.R
\name{uniformatValidate}
\alias{uniformatValidate}
\title{Validates a UNIFORMAT string.}
\usage{
uniformatValidate(uniformat)
}
\arguments{
\item{uniformat}{A string of HLA allele names and operators in the UNIFORMAT format signifying their relation with one another.}
}
\value{
If the input string is valid, the original string is returned. If the input string is invalid, an error message is generated.
}
\description{
A function that validates that a UNIFORMAT formatted string, ensuring that it does not contain any characters not supported in UNIFORMAT.
}
\examples{
uniformatValidate("A*02:01,A*03:01|A*02:01,A*03:02|A*02:02,A*03:01|A*02:02,A*03:02")

}
\references{
Nunes Tissue Antigens 2007;69(s1):203-205 https://doi.org/10.1111/j.1399-0039.2006.00808.x
}
