% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDNASequenceSearch.R
\name{unigDNAalign}
\alias{unigDNAalign}
\title{Generate a genomic alignment for specific HLA alleles at a set of positions}
\usage{
unigDNAalign(alleles, positions)
}
\arguments{
\item{alleles}{A vector of un-prefixed HLA locus names}

\item{positions}{a vector of nucleotide positions, against which all loci will be aligned.}
}
\value{
A data frame of allele names and the corresponding nucleotide sequence for specified position. a nucleotide sequence is not returned for a specific allele if input allele is not available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a nucleotide at the input position.
}
\description{
Generates a genomic (gDNA) nucleotide alignment at a single set of nucleotide positions for HLA alleles at one or more loci.
}
\note{
for internal use only
}
\examples{
\dontrun{
unigDNAalign("DQA1*01:01:01:01",c(32:58))
}
}
