% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDNASequenceSearch.R
\name{unigDNAsearch}
\alias{unigDNAsearch}
\title{Search genomic HLA nucleotide sequences at single specified position for a specified HLA allele}
\usage{
unigDNAsearch(locus, allele, position, prefix = TRUE, trimmed = FALSE)
}
\arguments{
\item{locus}{An HLA locus in the HLAalignments$gen object in the HLAtools.data package.}

\item{allele}{The name of an allele at 'locus'.}

\item{position}{A numeric value identifying a single nucleotide position. If more than one position is specified, results will be generated for the first position.}

\item{prefix}{A logical that indicates if the position number should be included in the result.}

\item{trimmed}{A logical identifying whether 'allele' is a two-field, "trimmed" allele name (trimmed=TRUE), or a full-length name (trimmed=FALSE). The default value is FALSE.}
}
\value{
The nucleotide sequence at the specified position for the specified allele. If sequence is not available for a position, no value will be returned.
}
\description{
Generates a character string of the genomic (gDNA) nucleotide sequence at the specified position for the specified HLA allele name.
}
\note{
For internal HLAtools use only.
}
\examples{
\dontrun{
unigDNAsearch("DRB1", "15:07", 57, trimmed=TRUE)
unigDNAsearch("DRB1", "15:07:01", 57)
}
}
