% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.r
\name{update_IMGTHLA}
\alias{update_IMGTHLA}
\title{Update a local IMGTHLA repo}
\usage{
update_IMGTHLA()
}
\value{
\code{invisible(NULL)}
}
\description{
Update a local IMGTHLA repo
}
\examples{
\dontrun{
update_IMGTHLA()
}
}

