% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildAlleleListHistory.R
\name{updateAlleleListHistory}
\alias{updateAlleleListHistory}
\title{Build the AlleleList History R Object}
\usage{
updateAlleleListHistory()
}
\value{
A list object containing two data frames. alleleListHistory$Version identifies the date and version of alleleListHistory$AlleleListHistory. alleleListHistory$AlleleListHistory relates each IPD-IMGT/HLA Database Accession ID (HLA_ID) to the corresponding HLA allele name in each release version since 1.05.0.
}
\description{
Consumes the ANHIG/IMGTHLA GitHub repository's allelelisthistory.txt file to create the AlleleListHistory object.
}
