% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{utr}
\alias{utr}
\title{Access UTR sequences}
\usage{
utr(x, utr = NULL, ...)
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} object.}

\item{utr}{The feature number or \code{NULL} (Retrieves all UTRs).}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{DNAStringSet}.
}
\description{
Access UTR sequences
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
utr(x, 1)
}
}
