% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNItoGLS.R
\name{validateUniformat}
\alias{validateUniformat}
\title{Validate a UNIFORMAT String}
\usage{
validateUniformat(uniformat)
}
\arguments{
\item{uniformat}{A character string of allele names and operators in the UNIFORMAT format.}
}
\value{
A logical. TRUE is returned when all characters in 'uniformat' are permitted. FALSE is returned when forbidden characters are present.
}
\description{
Evaluates a UNIFORMAT string to identify unsupported characters.
}
\examples{
validateUniformat("A*02:01,A*03:01|A*02:01,A*03:02|A*02:02,A*03:01|A*02:02,A*03:02")

}
\references{
Nunes Tissue Antigens 2007;69(s1):203-205 https://doi.org/10.1111/j.1399-0039.2006.00808.x
}
