% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchandBuildAlignments.R
\name{variantTable}
\alias{variantTable}
\title{Builds a Data Frame of Variant Information for Specific Positions in an Alignment}
\usage{
variantTable(alignType, locus, position)
}
\arguments{
\item{alignType}{The type of alignment being searched. Allowed values are "prot", codon", "nuc" and "gen".  Only one 'alignType' value is allowed.}

\item{locus}{A locus supported by the IPD-IMGT/HLA Database.}

\item{position}{A numeric variant position.}
}
\value{
a data frame of three columns identifying the variants, and the number and frequency of each in the alignment.
}
\description{
Builds a data frame sequence variants at specific positions in specified alignments in HLAalignments, along with the counts and frequencies for each variant.
}
\note{
For internal HLAtools use.
}
