% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{allele_id}
\alias{allele_id}
\title{Access allele IDs}
\usage{
allele_id(x, ...)
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} or \linkS4class{HLAAllele} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of allele IDs.
}
\description{
Access allele IDs
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
allele_id(x[is_complete(x)])
}
}
