% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{allele_name}
\alias{allele_name}
\title{Access allele names}
\usage{
allele_name(x, ...)
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} or \linkS4class{HLAAllele} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of allele names.
}
\description{
Access allele names
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
allele_name(x[is_complete(x)])
}
}
