% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alleleListHistory}
\alias{alleleListHistory}
\title{Allele Names Across All Release Versions}
\format{
A large list of 2 elements that identify the table of allele names for each IPD-IMGT/HLA reference Database release and the source release
\itemize{
\item(Version: the Date and IPD-IMGT HLA Release version under which the alleleListHistory object was made)
\item(AlleleListHistory: a dataframe version of the alleleListHistory.txt file)
}
}
\source{
https://raw.githubusercontent.com/ANHIG/IMGTHLA/Latest/Allelelist_history.txt
}
\usage{
data(alleleListHistory)
}
\description{
A large list object of two elements. The first element contains version information. The second element contains all releases of allele name lists in the IPD-IMGT/HLA Database.
This object is built by the UpdateAlleleListHistory() function.
}
\note{
The Allele List History file in IPD-IMGT/HLA Database release version 3.60.0 is too large to be included in the HLAtools package. The alleleListHistory object for release version 3.59.0 is included in the package. The alleleListHistory object in release versions 3.60.0 and later can be loaded via the updateAll() function.
}
\keyword{datasets}
