% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAlignments.R
\name{buildAlignments}
\alias{buildAlignments}
\title{Build Amino Acid, cDNA and gDNA Alignments}
\usage{
buildAlignments(loci, source, version = "Latest")
}
\arguments{
\item{loci}{A vector of HLA gene names (ex. "DRB1", c("A","C")).}

\item{source}{A vector of alignment types. The allowed values are "AA", "cDNA", and "gDNA". If 'source' is "AA", both amino acid and codon alignments are generated. If source is 'cDNA' or 'gDNA', a single cDNA or gDNA nucleotide alignment is generated. Up to four alignments will be returned for a locus, as determined by its ability to be transcribed or translated.}

\item{version}{The desired release version (branch) of the ANHIG/IMGTHLA Github repository (e.g. '3.53.0'). The default value ('Latest') returns alignments for the most recent branch.}
}
\value{
A list object with a data frame of all allele names (and trimmed allele names) and their corresponding sequences (Amino Acid, cDNA, or gDNA) for a specific loci, as well as version details for the returned information. Also returns locations of exons in relation to amino acids, cDNA, and gDNA.
}
\description{
Returns a list of data frames of amino-acid, cDNA and gDNA alignments, and version information.
}
\examples{
\dontrun{
buildAlignments(loci = "DRB1", source = "AA")
buildAlignments(loci = "DRB1", source = c("AA", "cDNA"))
}
}
