% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDNASequenceSearch.R
\name{cDNAsearch}
\alias{cDNAsearch}
\title{Search HLA alignments for specific alleles and codon positions.}
\usage{
cDNAsearch(allelename, positions, prefix = TRUE, sep = "~")
}
\arguments{
\item{allelename}{A full or 2-field HLA allele name, excluding the "HLA-" prefix.}

\item{positions}{A vector of nucleotide positions (e.g., c(-17,1,22,130)); indel positions are named using decimals. So the first indel between positions 26 and 27 is named 26.1, and the second indel between 26 and 27 is 26.2, etc.}

\item{prefix}{A logical that indicates if the position number(s) should be included in the result.}

\item{sep}{The value that will be used to separate the sequences for each position. No value can be specified (sep=""). If prefix=FALSE and sep="", a sequence block is returned.}
}
\value{
A character string containing the corresponding codon sequence for each position in 'positions' for 'allelename'. A codon sequence is not returned if 'allelename' is not not found in the pertinent alignment. A position will be empty if 'allelename' does not have a codon at the specified position.
}
\description{
Searches for alleles in the ANHIG/IMGT HLA database, searches for codons for submitted allele name(s) and position(s).
}
\examples{
\dontrun{
cDNAsearch("DRB1*15:07:01",11:22)
cDNAsearch("DRB1*15:07",11:22)
}
}
