% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSequences.R
\name{compareSequences}
\alias{compareSequences}
\title{Identify the differences between a pair of allele sequences at a locus}
\usage{
compareSequences(alignType, alleles)
}
\arguments{
\item{alignType}{The type of alignment being searched. Allowed values are "codon","gen", nuc" and "prot". Only one 'alignType' value is permitted.}

\item{alleles}{A vector containing two full-length allele names at a locus (ex. c("DPA1\emph{01:03:38:01","DPA1}01:03:38:02")).}
}
\value{
A data two-row frame identifying the positions and sequences at which the two alleles differ. Positions for which the sequence of either allele is unknown are ignored.
}
\description{
Compares the sequences of two alleles at a locus, and identifies the differences between them at specific positions
}
\examples{
\dontrun{
compareSequences(alignType = "gen", alleles = c("DPA1*01:03:38:01","DPA1*01:03:38:02"))
}
}
