% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{db_version}
\alias{db_version}
\title{Get IPD-IMGT/HLA database version}
\usage{
db_version(x, ...)
}
\arguments{
\item{x}{A \link{HLAGene} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \link{numeric_version} object.
}
\description{
Retrieve the \href{https://www.ebi.ac.uk/ipd/imgt/hla/}{IPD-IMGT/HLA} from which
a HLAgene object has been created.
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
db_version(x)
}
}
