% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{ethnicity}
\alias{ethnicity}
\title{Access ethnicity status}
\usage{
ethnicity(x, ...)
}
\arguments{
\item{x}{A \code{\link{HLAGene}} or \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of colon-separated sample ethnicities.
}
\description{
Access ethnicity status
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
ethnicity(x[is_complete(x)])
}
}
