% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{exon}
\alias{exon}
\title{Retrieve exon sequences}
\usage{
exon(x, exon = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{HLAGene}} object.}

\item{exon}{The feature number or \code{NULL} (Retrieves all exons).}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\linkS4class{DNAStringSet}}.
}
\description{
Retrieve exon sequences
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
exon(x, 2)
}
}
