% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{features}
\alias{features}
\alias{features<-}
\title{Access feature data}
\usage{
features(x, type = NULL, ...)

features(x, ...) <- value
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} or \linkS4class{HLAAllele} object.}

\item{type}{Which feature type.}

\item{...}{Further arguments passed to methods.}

\item{value}{A \code{HLAFeatureList} object.}
}
\value{
A \linkS4class{HLAFeatureList} object.
}
\description{
Access feature data
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
features(x[is_complete(x)])
}
}
