% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.r
\name{fetch_IMGTHLA}
\alias{fetch_IMGTHLA}
\alias{update_IMGTHLA}
\alias{check_IMGTHLA}
\title{Clone or pull from the \href{https://github.com/ANHIG/IMGTHLA}{ANHIG/IMGTHLA} repo}
\usage{
fetch_IMGTHLA(local_path = getOption("hlatools.local_repos"))

update_IMGTHLA(local_path = getOption("hlatools.local_repos"))

check_IMGTHLA(local_path = getOption("hlatools.local_repos"))
}
\arguments{
\item{local_path}{Local directory to clone to. Defaults to \file{~/local/db/IMGTHLA}.}
}
\value{
A \linkS4class{git_repository} object.

\code{invisible(NULL)}

A diagnostic message.
}
\description{
Clone or pull from the \href{https://github.com/ANHIG/IMGTHLA}{ANHIG/IMGTHLA} repo
}
\note{
Due to limitations in the \pkg{git2r} this will not work
behind a proxy.
}
\examples{
\dontrun{
check_IMGTHLA()
repo <- fetch_IMGTHLA()
update_IMGTHLA()
}
}
