% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.r
\name{fetch_IMGTHLA}
\alias{fetch_IMGTHLA}
\title{Clone the IMGTHLA repo on github}
\usage{
fetch_IMGTHLA(local_path = getOption("hlatools.local_repos"))
}
\arguments{
\item{local_path}{Local directory to clone to. Defaults to \file{~/local/db/IMGTHLA}.}
}
\value{
A S4 \code{\linkS4class{git_repository}} object.
}
\description{
Clone the IMGTHLA repo on github
}
\examples{
\dontrun{
repo <- fetch_IMGTHLA()
}
}

