% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_tables.r
\name{g_table}
\alias{g_table}
\title{Class: g_tbl}
\source{
\url{http://hla.alleles.org/nomenclature/g_groups.html}.
}
\usage{
g_table()
}
\value{
A table with the fields:
\itemize{
\item "gene": A HLA gene.
\item "code": The G Code.
\item "subtype": The '/'-separated subtypes into which a G Code expands.
}
}
\description{
Constructor for a <\code{g_tbl}> object.
}
\examples{
\dontrun{
g_codes <- g_table()
}
}
\seealso{
\code{\link{nmdp_table}}
}
