% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlleleTrim.R
\name{getField}
\alias{getField}
\title{Trims colon-delimited HLA allele names by field}
\usage{
getField(allele, res, append = FALSE)
}
\arguments{
\item{allele}{HLA allele.}

\item{res}{Resolution desired.}

\item{append}{A logical that, when TRUE, appends the expression variant suffix of a full-length allele name to a truncated allele name. The default value is FALSE.}
}
\description{
Trims a properly formatted colon-delimited HLA allele name to a desired number of fields.

If an allele name with an expression-variant suffix is truncated, the suffix can be appended to the end of the truncated allele name.
}
\note{
For internal HLAtools use.
}
\examples{
getField("HLA-A*01:01:01:01", 3)
getField("DRB1*11:01:01:12N", 2,TRUE)

}
