% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getFrame}
\alias{getFrame}
\title{Get feature reading frame}
\usage{
getFrame(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{HLARanges}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
An integer vector.
}
\description{
Get feature reading frame
}
\examples{
##
}
