% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{getId}
\alias{getId}
\title{Get feature id}
\usage{
getId(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{HLARanges}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector.
}
\description{
Get feature id
}
\examples{
##
}

