% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{getId}
\alias{getId}
\title{Get feature id}
\usage{
getId(x, ...)
}
\arguments{
\item{x}{A \linkS4class{HLARanges} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \link{character} vector.
}
\description{
Get feature id
}
\examples{
##
}
