% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionValidation.R
\name{getLatestVersion}
\alias{getLatestVersion}
\title{Identify the Latest IPD-IMGT/HLA Database Release}
\usage{
getLatestVersion()
}
\value{
A dot-delimited character string identifying the latest release version (branch) of the ANHIG/IMGTHLA Github repository.
}
\description{
Identifies the most recent version of the IPD-IMGT/HLA Database available in the ANHIG/IMGTHLA Github repository.
}
\note{
For internal HLAtools use.
}
\examples{
getLatestVersion()

}
