% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{getOrder}
\alias{getOrder}
\title{Access feature order}
\usage{
getOrder(x, ...)
}
\arguments{
\item{x}{A \linkS4class{HLARanges} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
An \link{integer} vector.
}
\description{
Access feature order
}
\examples{
##
}
