% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{getStatus}
\alias{getStatus}
\title{Access feature status}
\usage{
getStatus(x, ...)
}
\arguments{
\item{x}{A \linkS4class{HLARanges} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \link{character} vector.
}
\description{
Access feature status
}
\examples{
##
}
