% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLupdater.R
\name{GLV2}
\alias{GLV2}
\title{Format GL String Code version number.}
\usage{
GLV2(Version)
}
\arguments{
\item{Version}{A character string of the desired IPD-IMGT/HLA Database version, going back to version 1.05.0.}
}
\value{
Returns either a single, compressed character string-formatted release version, or a character vector of potential release versions.
}
\description{
Returns a compressed IPD-IMGT/HLA Database release version or a vector of dot-delimited release version options if the specified version is not present in the alleleListHistory object.
}
\note{
For internal use only.
}
\examples{
GLV2("3.34.0")
GLV2("3.0.0")

}
\references{
Mack et al. HLA 2023;102(2):206-212 https://doi.org/10.1111/tan.15126

Mack et al. HLA 05 July 2023 https://doi.org/10.1111/tan.15145
}
