% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAAllele.R
\docType{class}
\name{HLAAllele-class}
\alias{HLAAllele-class}
\alias{names,HLAAllele-method}
\alias{length,HLAAllele-method}
\alias{c,HLAAllele-method}
\alias{[,HLAAllele,numeric,missing-method}
\alias{[,HLAAllele,logical,missing-method}
\alias{[,HLAAllele,character,missing-method}
\title{Class `"HLAAllele"``}
\usage{
\S4method{names}{HLAAllele}(x)

\S4method{length}{HLAAllele}(x)

\S4method{c}{HLAAllele}(x, ..., recursive = FALSE)

\S4method{[}{HLAAllele,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{HLAAllele,logical,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{HLAAllele,character,missing}(x, i, j, ..., drop = TRUE)
}
\description{
A container for data parsed from the IPD-IMGT/HLA hla.xml file.
Part of a \link[=HLAGene_]{HLAGene} object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{names}: Get sequence names from \linkS4class{HLAAllele}s.

\item \code{length}: Get number of \linkS4class{HLAAllele}s.

\item \code{c}: Combine \linkS4class{HLAAllele} objects.

\item \code{[}: Subset \linkS4class{HLAAllele} objects.

\item \code{[}: Subset \linkS4class{HLAAllele} objects.

\item \code{[}: Subset \linkS4class{HLAAllele} objects.
}}

\section{Slots}{

\describe{
\item{\code{locus}}{A \code{character} string.}

\item{\code{sequence}}{A \linkS4class{DNAStringSet} object.}

\item{\code{metadata}}{A \linkS4class{DataFrame} object.}

\item{\code{features}}{A \linkS4class{HLARangesList} object.}
}}

\examples{
showClass("HLAAllele")
}
\seealso{
\link[=HLAGene_]{HLAGene}
}
\keyword{classes}
\keyword{internal}
