% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HlaAllele.r
\name{HLAAllele}
\alias{HLAAllele}
\title{Construct HLARanges objects}
\usage{
HLAAllele(node)
}
\arguments{
\item{node}{An allele node from a hla.xml object.}
}
\value{
A HLAAllele object
}
\description{
Construct HLARanges objects
}
\examples{
##
}

