% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAAllele.R
\name{HLAAllele}
\alias{HLAAllele}
\title{Constructor for \linkS4class{HLAAllele} objects}
\usage{
HLAAllele(nodes, locusname, ncores = parallel::detectCores())
}
\arguments{
\item{nodes}{Allele nodes from a \code{hla.xml} object.}

\item{locusname}{A valid HLA locus name.}

\item{ncores}{The number of compute cores to use.}
}
\value{
A \linkS4class{HLAAllele} object
}
\description{
Constructor for \linkS4class{HLAAllele} objects
}
\note{
Don't run directly. This function is called by \code{\link[=parse_hla_alleles]{parse_hla_alleles()}}.
}
\examples{
showClass("HLAAllele")
}
\seealso{
\code{\link[=parse_hla_alleles]{parse_hla_alleles()}}
}
