% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLAatlas}
\alias{HLAatlas}
\title{identifies the boundary positions of exons, introns and UTRs in the amino acid, cDNA and gDNA alignments in the HLAalignments data object.}
\format{
A list of 4 elements that include the gene features boundaries in each ANHIG/IMGTHLA sequence alingment.
\itemize{
\item(prot: peptide-alignment atlases)
\item(nuc: cDNA-alignment atlases)
\item(gen: gDNA-alignment atlases)
\item(version: The IPD-IMGT/HLA Database release version under which these data were generated)
}
}
\source{
https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments
}
\usage{
data(HLAatlas)
}
\description{
A list object of sub-lists of R dataframes (atlases) for each locus with a protein (prot), cDNA (nuc), and gDNA (gen) alignment. Each atlas identifies the position of the exon, intron or UTR boundaries in an alignment.
This object is built by the atlasFull() function.
}
\keyword{datasets}
