% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAGene.r
\name{HLAGene}
\alias{HLAGene}
\title{Constructor for \code{\link[=HLAGene_]{HLAGene}} objects.}
\usage{
HLAGene(locusname)
}
\arguments{
\item{locusname}{A valid HLA gene name.}
}
\value{
A \code{\link[=HLAGene_]{HLAGene}} object
}
\description{
Constructor for \code{\link[=HLAGene_]{HLAGene}} objects.
}
\examples{
\dontrun{
HLAGene("DPB1")
}
}

