% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAGene.r
\name{HLAGene}
\alias{HLAGene}
\title{Constructor for \code{\link[=HLAGene_]{HLAGene}} objects.}
\usage{
HLAGene(locusname, ...)
}
\arguments{
\item{locusname}{A valid HLA gene name.}

\item{...}{Passed on.}
}
\value{
A \code{\link[=HLAGene_]{HLAGene}} object
}
\description{
Constructor for \code{\link[=HLAGene_]{HLAGene}} objects.
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
db_version(x)
locusname(x)
length(x)

## extract alleles with complete sequences
x[is_complete(x)]

## extract all Common alleles
x[cwd_status(x) == "Common"]
}
}
