% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAGene.R
\docType{class}
\name{HLAGene_}
\alias{HLAGene_}
\title{Class \code{"HLAGene"}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HLAgene(locusname, db_version = "Latest", db_path = getOption("hlatools.local_repos"),
       ncores = parallel::detectCores() - 2, with_dist = FALSE)
}
\value{
Object of \link{R6Class} representing an HLA gene.
}
\description{
Class \code{"HLAGene"}
}
\section{Methods}{

\describe{
\item{\code{x$new(locusname, db_version = "Latest", db_path = getOption("hlatools.local_repos"),
  ncores = parallel::detectCores() - 2, with_dist = FALSE)}}{Create a HLAGene object.}
\item{\code{x$print()}}{Display the HLAGene object.}
\item{\code{x$get_hlatools_version()}}{The package version under which a HLAGene object was created.}
\item{\code{x$get_db_version()}}{The IPD-IMGT/HLA database version.}
\item{\code{x$get_locusname()}}{The name of the HLA locus.}
\item{\code{x$get_alleles(allele)}}{A HLAAllele object.}
\item{\code{x$has_distances()}}{Has a distance matrix between all alleles been calculated?}
\item{\code{x$calculate_exon_distance_matrix()}}{Calculate a distance matrix
between all alleles based on the available exon sequences.}
\item{\code{x$get_closest_complete_neighbor(allele, partially = TRUE)}}{Get
the full-length allele that is closest to the query allele, based on the
available exon sequences.}
\item{\code{x$get_reference_sequence(allele)}}{Get the reference sequence for
query allele extended to full length by filling up missing sequence information
with the closest complete neighbour.}
\item{\code{x$get_extended_reference_set(verbose = FALSE)}{Get the reference
  sequences for all allele extended to full length by filling up missing sequence
  information with the closest complete neighbours.}}
}
}

\keyword{data}
\keyword{internal}
