% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLARanges.r
\docType{class}
\name{HLARanges-class}
\alias{HLARanges-class}
\title{Class \code{"HLARanges"}}
\description{
Extends \linkS4class{GRanges}.
}
\section{Slots}{

\describe{
\item{\code{id}}{character.}

\item{\code{order}}{integer.}

\item{\code{type}}{character.}

\item{\code{status}}{character.}

\item{\code{frame}}{integer.}
}}

\examples{
showClass("HLARanges")
}
\seealso{
\code{\link[=parse_hla_alleles]{parse_hla_alleles()}}, \linkS4class{HLAAllele}
}
\keyword{classes}
\keyword{internal}
