% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLARanges.r
\docType{class}
\name{HLARangesList-class}
\alias{HLARangesList-class}
\title{Class \code{"HLARangesList"}}
\description{
List container for \linkS4class{HLARanges} objects.
}
\examples{
showClass("HLARangesList")
}
\seealso{
\code{\link[=parse_hla_alleles]{parse_hla_alleles()}}, \linkS4class{HLAAllele}, \linkS4class{HLARanges}
}
\keyword{classes}
