% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLARanges.r
\name{HLARangesList}
\alias{HLARangesList}
\title{Constructor for \code{\linkS4class{HLARangesList}} objects}
\usage{
HLARangesList(...)
}
\arguments{
\item{...}{HLARanges objects.}
}
\value{
A \code{\linkS4class{HLARangesList}} object
}
\description{
Constructor for \code{\linkS4class{HLARangesList}} objects
}
\examples{
showClass("HLARangesList")
}
\seealso{
\code{\link{parse_hla_alleles}}, \code{\link{HLAAllele}},
\code{\link{HLARanges}}
}

