% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLARanges.r
\name{HLARangesList}
\alias{HLARangesList}
\title{Constructor for \linkS4class{HLARangesList} objects}
\usage{
HLARangesList(...)
}
\arguments{
\item{...}{\linkS4class{HLARanges} objects.}
}
\value{
A \linkS4class{HLARangesList} object
}
\description{
Constructor for \linkS4class{HLARangesList} objects
}
\examples{
showClass("HLARangesList")
}
\seealso{
\code{\link[=parse_hla_alleles]{parse_hla_alleles()}}, \linkS4class{HLAAllele}, \linkS4class{HLARanges}
}
