% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlatools-package.r
\docType{package}
\name{hlatools-package}
\alias{hlatools}
\alias{hlatools-package}
\title{A collection of tools to work with IPD-IMGT/HLA data}
\description{
A collection of tools to work with IPD-IMGT/HLA data
}
\details{
Details
}
\author{
\strong{Maintainer}: Gerhard Schöfl \email{gerhard.schofl@gmail.com}

}
