% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlatools-package.R
\docType{package}
\name{hlatools-package}
\alias{hlatools}
\alias{hlatools-package}
\title{A collection of tools to work with \href{https://www.ebi.ac.uk/ipd/imgt/hla/}{IPD-IMGT/HLA} data}
\description{
A collection of tools to work with IPD-IMGT/HLA data.
}
\details{
Details
}
\author{
\strong{Maintainer}: Gerhard Schöfl \email{schoefl@dkms-lab.de}

}
