% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{hlatools_version}
\alias{hlatools_version}
\title{Get hlatools package version}
\usage{
hlatools_version(x, ...)
}
\arguments{
\item{x}{A \link{HLAGene} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A \link{package_version} object.
}
\description{
Retrieve the package version under which a HLAgene object has been created.
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
hlatools_version(x)
}
}
