% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{intron}
\alias{intron}
\title{Access intron sequences}
\usage{
intron(x, intron = NULL, ...)
}
\arguments{
\item{x}{A \link[=HLAGene_]{HLAGene} object.}

\item{intron}{The feature number or \code{NULL} (Retrieves all introns).}

\item{...}{Further arguments passed to methods.}
}
\value{
A \linkS4class{DNAStringSet}.
}
\description{
Access intron sequences
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
intron(x, 1:2)
}
}
