% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{is_complete}
\alias{is_complete}
\title{Access completeness status}
\usage{
is_complete(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A logical vector.
}
\description{
Access completeness status
}
\examples{
\dontrun{
dpb1 <- parse_hla_alleles(read_hla_xml(), "DPB1")
dpb1[is_complete(dpb1), ]
}
}

