% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{locusname}
\alias{locusname}
\title{Access locus name}
\usage{
locusname(x, ...)
}
\arguments{
\item{x}{A \code{\link{HLAGene}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector.
}
\description{
Access locus name
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
locusname(x)
}
}
