% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locusValidator.R
\name{multiLocusValidation}
\alias{multiLocusValidation}
\title{Apply validateLocus() to multiple individual loci}
\usage{
multiLocusValidation(loci, verbose = TRUE)
}
\arguments{
\item{loci}{A vector of locus names.}

\item{verbose}{A logiccal value. If verbose = TRUE, messages describing invalid locus names are generated. If verbose = FALSE, no messages are generated.}
}
\value{
A vector of locus names that are present in HLAgazeteer$gen.
}
\description{
Applies validateLocus() to a vector of locus names, validates them against HLAgazeteer$gen, and returns a vector of validated locus names.
}
\note{
The results of this check should only be considered valid for the IPD-IMGT/HLA Database release version of the current HLAgazeteer.
}
\examples{
multiLocusValidation(loci = c("DRB1","DPB1","DQB8"))
multiLocusValidation(loci = c("A","B","C","D","Q"))

}
