% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryFunctions.R
\name{numFields}
\alias{numFields}
\title{Identify the Number of Fields in a Colon-Delimited Allele Name}
\usage{
numFields(allele)
}
\arguments{
\item{allele}{A character string of a colon-delimited HLA allele name.}
}
\value{
A numeric value describing the number of fields in the allele name.
}
\description{
Returns the number of fields in a colon-delimited HLA allele name. A value of 1 is returned for digit-delimited HLA allele names.
}
\examples{
numFields("HLA-A*01:01")
numFields("DRB1*04:03:01")
numFields("13:02:01:01")
}
