% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.r
\name{read_hla_xml}
\alias{read_hla_xml}
\alias{update_hla_xml}
\title{Fetch and parse or update the IPD-IMGT/HLA hla.xml file}
\usage{
read_hla_xml(remote = FALSE)

update_hla_xml()
}
\arguments{
\item{remote}{<\link{logical}>; if \code{TRUE} pull data from the IPD-IMGT/HLA ftp server,
if \code{FALSE} retrieve data from \code{getOption("hlatools.local_repos")}}
}
\value{
An \link[xml2:read_xml]{xml_document}.

\code{invisible(NULL)}
}
\description{
Fetch and parse or update the IPD-IMGT/HLA hla.xml file
}
\examples{
\dontrun{
doc <- read_hla_xml(remote = TRUE)
update_hla_xml()
}
}
