% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLupdater.R
\name{redec}
\alias{redec}
\title{Reintroduce version decimals.}
\usage{
redec(Cname)
}
\arguments{
\item{Cname}{A character string describing a version name in AlleleListHistory column format.}
}
\value{
A character string describing a version name with decimals in appropriate places according to the IPD-IMGT/HLA Database.
}
\description{
Correctly put decimals back into an AlleleListHistory column-formatted version name.
}
\note{
For internal use only.
}
\examples{
redec("X3090")

}
