% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{sequences}
\alias{sequences}
\alias{sequences<-}
\title{Get or set sequence data}
\usage{
sequences(x, ...)

sequences(x, ...) <- value
}
\arguments{
\item{x}{A \code{\link{HLAGene}} or \code{\linkS4class{HLAAllele}} object.}

\item{...}{Further arguments passed to methods.}

\item{value}{A \code{DNAStringSet} object.}
}
\value{
A DNAStringSet object.
}
\description{
Get or set sequence data
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
sequences(x[is_complete(x)])
}
}

