% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDNASequenceSearch.R,
%   R/cDNASequenceSearch_2_SM.R
\name{unicDNAalign}
\alias{unicDNAalign}
\title{Generate a codon or nucleotide alignment for specific alleles at specific positions.}
\usage{
unicDNAalign(alignType, alleles, positions)

unicDNAalign(alignType, alleles, positions)
}
\arguments{
\item{alignType}{The type of alignment being searched. Allowed values are "codon" and "nuc".}

\item{alleles}{A vector of un-prefixed HLA locus names.}

\item{positions}{A vector of codon positions, against which all loci will be aligned.}
}
\value{
A data frame of allele names and the corresponding codon sequence for specified position. a codon sequence is not returned for a specific allele if input allele is not available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a codon at the input position.

A data frame of allele names and the corresponding codon sequence for specified position. a codon sequence is not returned for a specific allele if input allele is not available in the ANHIG/IMGTHLA Github Repository. position will be left empty if specific allele does not have a codon at the input position.
}
\description{
Generates a codon or nucleotide cDNA alignment at a single set of codon positions for HLA alleles at one or more loci.

Generates a codon or nucleotide cDNA alignment at a single set of codon positions for HLA alleles at one or more loci.
}
\note{
For internal HLAtools use only.

For internal HLAtools use only.
}
\examples{
\dontrun{
unicDNAalign("DQA1*01:01:01:01",c(32:58))
}
\dontrun{
unicDNAalign("DQA1*01:01:01:01",c(32:58))
}
}
