% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateAll.R
\name{updateAll}
\alias{updateAll}
\title{Updates all of the data objects derived from IPD-IMGT/HLA sources.}
\usage{
updateAll(version = getLatestVersion())
}
\arguments{
\item{version}{The version of the ANHIG/IMGTHLA Github repository to build alignments from. The default value is to call the getLatestVersion() function to generate alignments for the most recent IPD-IMGT/HLA Database release.}
}
\value{
A list object containing data frames of protein (prot), coding nucleotide (nuc), and genomic nucleotide (gen) alignments for all genes in the specified IPD-IMGT/HLA Database release, along with the pertinent reference database release.
}
\description{
Applies extractGeneTypes(), buildGazeteer(), ffN(), atlasFull() and alignmentsFull() to build a set of alignments for all loci supported in the ANHIG/IMGTHLA GitHub repository.
}
\note{
For internal HLAtools use.
}
\examples{
\dontrun{
updateAll()
}

}
