% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateAll.R
\name{updateAll}
\alias{updateAll}
\title{Updates all of the package data objects derived from IPD-IMGT/HLA Database resources.}
\usage{
updateAll(updateType = "all", version = getLatestVersion())
}
\arguments{
\item{updateType}{A vector of the data objects to be updated. By default, updateAll() builds all five data objects (data="all"). Alternatively specific data objects can be updated; e.g., data="alleleListHistory" or data=c("alleleListHistory","fragmentFeatureNames").}

\item{version}{The version of the ANHIG/IMGTHLA Github repository to build these objects from. By default, updateAll() calls the getLatestVersion() function to identify the most recent IPD-IMGT/HLA Database release.}
}
\description{
Applies updateAlleleListHistory(), atlasFull(), buildGazeteer(), extractGeneTypes(), and ffN() and to update the alleleListHistory, HLAatlas, HLAgazeteer, IMGTHLAGeneTypes and fragmentFeatureNames data objects.
}
\details{
A new alleleListHistory data object will likely need to be generated with each IPD-IMGT/HLA Database release. The other data objects will likely only change if the IPD-IMGT/HLA Database includes additional genes.
}
\note{
Generating a new HLAatlas can take 5 minutes or more to complete.
}
\examples{
\dontrun{
updateAll()
}

}
