% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.r
\name{utr}
\alias{utr}
\title{Retrieve UTR sequences}
\usage{
utr(x, utr = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{HLAGene}} object.}

\item{utr}{The feature number or \code{NULL} (Retrieves all UTRs).}

\item{...}{Further arguments passed to methods.}
}
\value{
A \code{\linkS4class{DNAStringSet}}.
}
\description{
Retrieve UTR sequences
}
\examples{
\dontrun{
x <- HLAGene("DPB1")
utr(x, 1)
}
}
