% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryFunctions.R
\name{validateAllele}
\alias{validateAllele}
\title{Determine if an allele name is properly formed and present in HLAtools.data::HLAalignments}
\usage{
validateAllele(allele)
}
\arguments{
\item{allele}{A colon-delimited HLA allele name.}
}
\value{
A logical identifying if the allele name is present in the alignments (TRUE) or, if it is not in the alignments or is not valid not (FALSE).
}
\description{
Returns TRUE if an allele name is found in HLAtools.data::HLAalignments in either the 'allele' column of full-length allele names or the 'trimmed_allele' column of two-field allele names. Returns FALSE if the allele name is not properly formed, or if the allele name is not found in HLAtools.data::HLAalignments.
}
\note{
Messages will be returned to the console if the allele name is malformed, or the locus is invalid.
}
\examples{
\dontrun{
validateAllele("A*01:01:01:117")
validateAllele("A*01:01:01")
validateAllele("A*01:01")
}
}
